% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\docType{class}
\name{AggregateAnalysisBlock-class}
\alias{AggregateAnalysisBlock-class}
\title{Block that holds only ggplot data and ggplot that has been gathered from
other blocks and aggregated
This is the defaul concrete version that can be envoked if there is no
specific aggregator for a given analysis block class.}
\description{
Inherits from "VirtualAggregateAnalysisBlock",
}

