% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\name{aggregateAnalysisData}
\alias{aggregateAnalysisData}
\title{Aggregate data into the block}
\usage{
aggregateAnalysisData(object, data, with_id)
}
\arguments{
\item{object}{object of class "VirtualAggregateAnalysisBlock"}

\item{object}{data of class "data.frame"}

\item{object}{with_id of class "data.frame"}
}
\value{
\code{object} object of class "VirtualAggregateAnalysisBlock"
}
\description{
Aggregate data into the block
}

