% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trades.r
\docType{class}
\name{AggregateExtendedTradesAnalysisBlock-class}
\alias{AggregateExtendedTradesAnalysisBlock-class}
\title{Analysis Module for aggregation of Extended Trades}
\description{
Inherits from "AggregateAnalysisBlock"
}

