% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades.r
\docType{class}
\name{AverageDownTradesAnalysisBlock-class}
\alias{AverageDownTradesAnalysisBlock-class}
\title{Analysis Module for extraction of Average Down Trades}
\description{
Computation block class to pull data required for extraction of
average down trades.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualPositionDataHandler",
              "VirtualTradeDataHandler"
}

