% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades.r
\docType{class}
\name{AverageDownTradesData-class}
\alias{AverageDownTradesData-class}
\title{Extended Trades Reference Data class.}
\description{
Concrete S4 class storing data of Extended Trades info.
Generated only by AverageDownTradesAnalysisBlock module.
}
\details{
Inherits from "VirtualTradeData"
}

