% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades_focus.r
\docType{class}
\name{AverageDownTradesFocusAnalysisBlock-class}
\alias{AverageDownTradesFocusAnalysisBlock-class}
\title{Analysis Module for extraction of Average Down Trades}
\description{
Computation block class compute summary focus data for
average down trades.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler"
}

