% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buys_and_sells.r
\docType{class}
\name{BuysAndSellsAnalysisBlock-class}
\alias{BuysAndSellsAnalysisBlock-class}
\title{Number of extended buys and sells in month compared to number market down days}
\description{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler",
              "VirtualMarketDataHandler"
}

