% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades_focus.r
\docType{methods}
\name{dataRequest,AverageDownTradesFocusAnalysisBlock,data.frame-method}
\alias{dataRequest,AverageDownTradesFocusAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{AverageDownTradesFocusAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'AverageDownTradesFocusAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'AverageDownTradesFocusAnalysisBlock'.
}
\description{
Request data from data source
}

