% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buys_and_sells.r
\docType{methods}
\name{dataRequest,BuysAndSellsAnalysisBlock,data.frame-method}
\alias{dataRequest,BuysAndSellsAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{BuysAndSellsAnalysisBlock,data.frame}(object, key_values)
}
\arguments{
\item{object}{object of class 'BuysAndSellsAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'BuysAndSellsAnalysisBlock'.
}
\description{
Request data from data source
}

