% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns.r
\docType{methods}
\name{dataRequest,ImpliedFactorReturnsAnalysisBlock,data.frame-method}
\alias{dataRequest,ImpliedFactorReturnsAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{ImpliedFactorReturnsAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'ImpliedFactorReturnsAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'ImpliedFactorReturnsAnalysisBlock'.
}
\description{
Request data from data source
}

