% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_state.r
\docType{methods}
\name{dataRequest,ImpliedFactorReturnsStateAnalysisBlock,data.frame-method}
\alias{dataRequest,ImpliedFactorReturnsStateAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{ImpliedFactorReturnsStateAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'ImpliedFactorReturnsStateAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'ImpliedFactorReturnsStateAnalysisBlock'.
}
\description{
Request data from data source
}

