% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_return.r
\docType{methods}
\name{dataRequest,MarketReturnAnalysisBlock,data.frame-method}
\alias{dataRequest,MarketReturnAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{MarketReturnAnalysisBlock,data.frame}(object, key_values)
}
\arguments{
\item{object}{object of class 'MarketReturnAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'MarketReturnAnalysisBlock'.
}
\description{
Request data from data source
}

