% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_timeseries.r
\docType{methods}
\name{dataRequest,MarketStyleAnalysisBlock,data.frame-method}
\alias{dataRequest,MarketStyleAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{MarketStyleAnalysisBlock,data.frame}(object, key_values)
}
\arguments{
\item{object}{object of class 'MarketStyleAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'MarketStyleAnalysisBlock'.
}
\description{
Request data from data source
}

