% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_factor_stat.r
\docType{methods}
\name{dataRequest,MarketStyleFactorStatisticAnalysisBlock,data.frame-method}
\alias{dataRequest,MarketStyleFactorStatisticAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{MarketStyleFactorStatisticAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'MarketStyleFactorStatisticAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'MarketStyleFactorStatisticAnalysisBlock'.
}
\description{
Request data from data source
}

