% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions.r
\docType{methods}
\name{dataRequest,OffsidePositionsAnalysisBlock,data.frame-method}
\alias{dataRequest,OffsidePositionsAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{OffsidePositionsAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'OffsidePositionsAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'OffsidePositionsAnalysisBlock'.
}
\description{
Request data from data source
}

