% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_cumulative_pnl.r
\docType{methods}
\name{dataRequest,OffsidePositionsCumulativePnLAnalysisBlock,data.frame-method}
\alias{dataRequest,OffsidePositionsCumulativePnLAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{OffsidePositionsCumulativePnLAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'OffsidePositionsCumulativePnLAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'OffsidePositionsCumulativePnLAnalysisBlock'.
}
\description{
Request data from data source
}

