% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl_traded_in_long_short_hedge.r
\docType{methods}
\name{dataRequest,PnLTradedInLongShortHedgeAnalysisBlock,data.frame-method}
\alias{dataRequest,PnLTradedInLongShortHedgeAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{PnLTradedInLongShortHedgeAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PnLTradedInLongShortHedgeAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PnLTradedInLongShortHedgeAnalysisBlock'.
}
\description{
Request data from data source
}

