% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_exposures.r
\docType{methods}
\name{dataRequest,PortfolioCoreFactorExposuresAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioCoreFactorExposuresAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{PortfolioCoreFactorExposuresAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PortfolioCoreFactorExposuresAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PortfolioCoreFactorExposuresAnalysisBlock'.
}
\description{
Request data from data source
}

