% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_hedge_returns.r
\docType{methods}
\name{dataRequest,PortfolioCoreFactorHedgeReturnsAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioCoreFactorHedgeReturnsAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{PortfolioCoreFactorHedgeReturnsAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PortfolioCoreFactorHedgeReturnsAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PortfolioCoreFactorHedgeReturnsAnalysisBlock'.
}
\description{
Request data from data source
}

