% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_returns.r
\docType{methods}
\name{dataRequest,PortfolioCoreFactorReturnsAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioCoreFactorReturnsAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{PortfolioCoreFactorReturnsAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PortfolioCoreFactorReturnsAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PortfolioCoreFactorReturnsAnalysisBlock'.
}
\description{
Request data from data source
}

