% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_exposure.r
\docType{methods}
\name{dataRequest,PortfolioFactorExposureAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioFactorExposureAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{PortfolioFactorExposureAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PortfolioFactorExposureAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PortfolioFactorExposureAnalysisBlock'.
}
\description{
Request data from data source
}

