% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_exposures.r
\docType{methods}
\name{dataRequest,PortfolioFactorExposuresAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioFactorExposuresAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{PortfolioFactorExposuresAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PortfolioFactorExposuresAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PortfolioFactorExposuresAnalysisBlock'.
}
\description{
Request data from data source
}

