% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_returns.r
\docType{methods}
\name{dataRequest,PortfolioFactorReturnsAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioFactorReturnsAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{PortfolioFactorReturnsAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PortfolioFactorReturnsAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PortfolioFactorReturnsAnalysisBlock'.
}
\description{
Request data from data source
}

