% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{dataRequest,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock'.
}
\description{
Request data from data source
}

