% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk.r
\docType{methods}
\name{dataRequest,PortfolioInstrumentMCTRAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioInstrumentMCTRAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{PortfolioInstrumentMCTRAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PortfolioInstrumentMCTRAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PortfolioInstrumentMCTRAnalysisBlock'.
}
\description{
Request data from data source
}

