% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{methods}
\name{dataRequest,PortfolioVarianceDecompositionAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioVarianceDecompositionAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{PortfolioVarianceDecompositionAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PortfolioVarianceDecompositionAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PortfolioVarianceDecompositionAnalysisBlock'.
}
\description{
Request data from data source
}

