% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_revisits.r
\docType{methods}
\name{dataRequest,PositionRevisitsAnalysisBlock,data.frame-method}
\alias{dataRequest,PositionRevisitsAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{PositionRevisitsAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PositionRevisitsAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PositionRevisitsAnalysisBlock'.
}
\description{
Request data from data source
}

