% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_revisits_delta_prev_quarter.r
\docType{methods}
\name{dataRequest,PositionRevisitsDeltaPrevQuarterAnalysisBlock,data.frame-method}
\alias{dataRequest,PositionRevisitsDeltaPrevQuarterAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{PositionRevisitsDeltaPrevQuarterAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PositionRevisitsDeltaPrevQuarterAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PositionRevisitsDeltaPrevQuarterAnalysisBlock'.
}
\description{
Request data from data source
}

