% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_capital_distribution.r
\docType{methods}
\name{dataRequest,PositionsHoldingCapitalDistributionAnalysisBlock,data.frame-method}
\alias{dataRequest,PositionsHoldingCapitalDistributionAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{PositionsHoldingCapitalDistributionAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PositionsHoldingCapitalDistributionAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PositionsHoldingCapitalDistributionAnalysisBlock'.
}
\description{
Request data from data source
}

