% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_day_zero_pnl.r
\docType{methods}
\name{dataRequest,PositionsHoldingDayZeroPnLAnalysisBlock,data.frame-method}
\alias{dataRequest,PositionsHoldingDayZeroPnLAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{PositionsHoldingDayZeroPnLAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PositionsHoldingDayZeroPnLAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PositionsHoldingDayZeroPnLAnalysisBlock'.
}
\description{
Request data from data source
}

