% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown.r
\docType{methods}
\name{dataRequest,StrategyBreakdownAnalysisBlock,data.frame-method}
\alias{dataRequest,StrategyBreakdownAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{StrategyBreakdownAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'StrategyBreakdownAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'StrategyBreakdownAnalysisBlock'.
}
\description{
Request data from data source
}

