% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_signal_characteristic_and_effeciveness.r
\docType{methods}
\name{Process,StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock-method}
\alias{Process,StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{

  \S4method{Process}{StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock"}
}
\value{
\code{object} object object of class "StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

