% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_value_traded_per_signal.r
\docType{methods}
\name{dataRequest,StrategyBreakdownValueTradedPerSignalAnalysisBlock,data.frame-method}
\alias{dataRequest,StrategyBreakdownValueTradedPerSignalAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{StrategyBreakdownValueTradedPerSignalAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'StrategyBreakdownValueTradedPerSignalAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'StrategyBreakdownValueTradedPerSignalAnalysisBlock'.
}
\description{
Request data from data source
}

