% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades_pnl_outof_on_results_day.r
\docType{methods}
\name{dataRequest,TradesPnLOutOfOnResultsDayAnalysisBlock,data.frame-method}
\alias{dataRequest,TradesPnLOutOfOnResultsDayAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{TradesPnLOutOfOnResultsDayAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'TradesPnLOutOfOnResultsDayAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'TradesPnLOutOfOnResultsDayAnalysisBlock'.
}
\description{
Request data from data source
}

