% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_traded_in_long_short_hedge.r
\docType{methods}
\name{dataRequest,ValueTradedInLongShortHedgeAnalysisBlock,data.frame-method}
\alias{dataRequest,ValueTradedInLongShortHedgeAnalysisBlock,data.frame-method}
\title{Request data from data source}
\usage{

  \S4method{dataRequest}{ValueTradedInLongShortHedgeAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'ValueTradedInLongShortHedgeAnalysisBlock'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'ValueTradedInLongShortHedgeAnalysisBlock'.
}
\description{
Request data from data source
}

