% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\docType{methods}
\name{dataRequest,VirtualAggregateAnalysisBlock,data.frame-method}
\alias{dataRequest,VirtualAggregateAnalysisBlock,data.frame-method}
\title{Request data from data source: Overriden to have no function because
this does not make sense in the context of aggregate blocks}
\usage{
\S4method{dataRequest}{VirtualAggregateAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class "VirtualAggregateAnalysisBlock".}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'VirtualAggregateAnalysisBlock'.
}
\description{
Request data from data source: Overriden to have no function because
this does not make sense in the context of aggregate blocks
}

