% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/offside_positions.r, R/average_down_trades.r, R/average_down_trades_focus.r, R/extended_trades.r, R/buys_and_sells.r, R/implied_factor_returns_state.r, R/market_return.r, R/portfolio_variance_decomposition.r, R/market_style_factor_stat.r, R/market_style_timeseries.r, R/offside_positions_cumulative_pnl.r, R/value_traded_in_long_short_hedge.r, R/pnl_traded_in_long_short_hedge.r, R/portfolio_core_factor_exposures.r, R/portfolio_core_factor_hedge_exposures.r, R/portfolio_core_factor_hedge_returns.r, R/portfolio_core_factor_returns.r, R/portfolio_factor_exposure.r, R/portfolio_factor_exposures.r, R/portfolio_factor_returns.r, R/portfolio_factor_state_prob_cond_loss.r, R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r, R/portfolio_instrument_marginal_contribution_to_risk.r, R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r, R/position_revisits.r, R/position_revisits_delta_prev_quarter.r, R/positions_holding_capital_distribution.r, R/positions_holding_day_zero_pnl.r, R/positions_holding_period.r, R/strategy_breakdown.r, R/strategy_breakdown_value_traded_per_signal.r, R/trades_performance_on_results_day.r, R/trades_pnl_outof_on_results_day.r
\docType{methods}
\name{dataRequest}
\alias{dataRequest}
\alias{dataRequest,AverageDownTradesAnalysisBlock,data.frame-method}
\alias{dataRequest,AverageDownTradesFocusAnalysisBlock,data.frame-method}
\alias{dataRequest,BuysAndSellsAnalysisBlock,data.frame-method}
\alias{dataRequest,ExtendedTradesAnalysisBlock,data.frame-method}
\alias{dataRequest,ImpliedFactorReturnsStateAnalysisBlock,data.frame-method}
\alias{dataRequest,MarketReturnAnalysisBlock,data.frame-method}
\alias{dataRequest,MarketStyleAnalysisBlock,data.frame-method}
\alias{dataRequest,MarketStyleFactorStatisticAnalysisBlock,data.frame-method}
\alias{dataRequest,OffsidePositionsAnalysisBlock,data.frame-method}
\alias{dataRequest,OffsidePositionsCumulativePnLAnalysisBlock,data.frame-method}
\alias{dataRequest,PnLTradedInLongShortHedgeAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioCoreFactorExposuresAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioCoreFactorHedgeExposuresAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioCoreFactorHedgeReturnsAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioCoreFactorReturnsAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioFactorExposureAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioFactorExposuresAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioFactorReturnsAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioFactorStateProbCondLossAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioInstrumentMCTRAnalysisBlock,data.frame-method}
\alias{dataRequest,PortfolioVarianceDecompositionAnalysisBlock,data.frame-method}
\alias{dataRequest,PositionRevisitsAnalysisBlock,data.frame-method}
\alias{dataRequest,PositionRevisitsDeltaPrevQuarterAnalysisBlock,data.frame-method}
\alias{dataRequest,PositionsHoldingCapitalDistributionAnalysisBlock,data.frame-method}
\alias{dataRequest,PositionsHoldingDayZeroPnLAnalysisBlock,data.frame-method}
\alias{dataRequest,PositionsHoldingPeriodAnalysisBlock,data.frame-method}
\alias{dataRequest,StrategyBreakdownAnalysisBlock,data.frame-method}
\alias{dataRequest,StrategyBreakdownValueTradedPerSignalAnalysisBlock,data.frame-method}
\alias{dataRequest,TradesPerformanceOnResultsDayAnalysisBlock,data.frame-method}
\alias{dataRequest,TradesPnLOutOfOnResultsDayAnalysisBlock,data.frame-method}
\alias{dataRequest,ValueTradedInLongShortHedgeAnalysisBlock,data.frame-method}
\alias{dataRequest,VirtualAnalysisBlock,data.frame-method}
\title{Request data from data sources}
\usage{
\S4method{dataRequest}{OffsidePositionsAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{AverageDownTradesAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{AverageDownTradesFocusAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{ExtendedTradesAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{BuysAndSellsAnalysisBlock,data.frame}(object, key_values)


  \S4method{dataRequest}{ImpliedFactorReturnsStateAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{MarketReturnAnalysisBlock,data.frame}(object, key_values)


  \S4method{dataRequest}{PortfolioVarianceDecompositionAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{MarketStyleFactorStatisticAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{MarketStyleAnalysisBlock,data.frame}(object, key_values)


  \S4method{dataRequest}{OffsidePositionsCumulativePnLAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{ValueTradedInLongShortHedgeAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PnLTradedInLongShortHedgeAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PortfolioCoreFactorExposuresAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PortfolioCoreFactorHedgeExposuresAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PortfolioCoreFactorHedgeReturnsAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PortfolioCoreFactorReturnsAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{PortfolioFactorExposureAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PortfolioFactorExposuresAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{PortfolioFactorReturnsAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PortfolioFactorStateProbCondLossAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{PortfolioInstrumentMCTRAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{PositionRevisitsAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PositionRevisitsDeltaPrevQuarterAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PositionsHoldingCapitalDistributionAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{PositionsHoldingDayZeroPnLAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{PositionsHoldingPeriodAnalysisBlock,data.frame}(object,
  key_values)

\S4method{dataRequest}{StrategyBreakdownAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{StrategyBreakdownValueTradedPerSignalAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{TradesPerformanceOnResultsDayAnalysisBlock,data.frame}(object,
  key_values)


  \S4method{dataRequest}{TradesPnLOutOfOnResultsDayAnalysisBlock,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class derived from"VirtualAnalysisBlock".}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class derived from 'VirtualAnalysisBlock'.

\code{object} object of class 'TradesPerformanceOnResultsDayAnalysisBlock'.

\code{object} object of class 'TradesPnLOutOfOnResultsDayAnalysisBlock'.
}
\description{
Request data from data sources
}
\section{Methods (by class)}{
\itemize{
\item \code{object = OffsidePositionsAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = AverageDownTradesAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = AverageDownTradesFocusAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = ExtendedTradesAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = BuysAndSellsAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = ImpliedFactorReturnsStateAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = MarketReturnAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioVarianceDecompositionAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = MarketStyleFactorStatisticAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = MarketStyleAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = OffsidePositionsCumulativePnLAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = ValueTradedInLongShortHedgeAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PnLTradedInLongShortHedgeAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioCoreFactorExposuresAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioCoreFactorHedgeExposuresAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioCoreFactorHedgeReturnsAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioCoreFactorReturnsAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioFactorExposureAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioFactorExposuresAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioFactorReturnsAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioFactorStateProbCondLossAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioInstrumentMCTRAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PositionRevisitsAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PositionRevisitsDeltaPrevQuarterAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PositionsHoldingCapitalDistributionAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PositionsHoldingDayZeroPnLAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = PositionsHoldingPeriodAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = StrategyBreakdownAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = StrategyBreakdownValueTradedPerSignalAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = TradesPerformanceOnResultsDayAnalysisBlock,key_values = data.frame}: Request data from data source

\item \code{object = TradesPnLOutOfOnResultsDayAnalysisBlock,key_values = data.frame}: Request data from data source
}}

