% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trades.r
\docType{class}
\name{ExtendedTradesAnalysisBlock-class}
\alias{ExtendedTradesAnalysisBlock-class}
\title{Analysis Module for extraction of Extended Trades}
\description{
Computation block class compute summary focus data for
extended trades.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler"
}

