% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trades_summary.r
\docType{class}
\name{ExtendedTradesSummaryAnalysisBlock-class}
\alias{ExtendedTradesSummaryAnalysisBlock-class}
\title{Analysis Module for computation extended trades summary}
\description{
Computation block class compute summary focus data for
extended trades.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler",
              "VirtualExtendedTradeDataHandler"
}

