% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\name{getOutputFrontendData}
\alias{getOutputFrontendData}
\title{Returns generated output frontend data}
\usage{
getOutputFrontendData(object, ...)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}
}
\value{
\code{frontend_data} object of class "data.frame"
}
\description{
Returns data to be used by any frontend using the module
}

