% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\name{getOutputGGPlot}
\alias{getOutputGGPlot}
\title{Returns generated output ggplot}
\usage{
getOutputGGPlot(object, ...)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}
}
\value{
\code{ggplot} object of class "ggplot"
}
\description{
Returns ggplot object generated as a result of analysis computation.
}

