% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\docType{methods}
\name{getOutputGGPlotData}
\alias{getOutputGGPlotData}
\alias{getOutputGGPlotData,VirtualAnalysisBlock-method}
\title{Returns generated output ggplot data}
\usage{
getOutputGGPlotData(object)

\S4method{getOutputGGPlotData}{VirtualAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}
}
\value{
\code{ggplot_data} object of class "data.frame"
}
\description{
Returns "data.frame" with data used to generate ggplot
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualAnalysisBlock}: Returns generated output ggplot data

Returns "data.frame" with data used to generate ggplot
}}

