% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\name{getOutputObject}
\alias{getOutputObject}
\title{Returns generated output object}
\usage{
getOutputObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}
}
\value{
\code{output} object of class derived from "VirtualReferenceData"
}
\description{
Returns Reference Data Object that possibly was computed by
analysis object. If module doesn't compute any output
returns NULL
}

