% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns.r
\docType{class}
\name{ImpliedFactorReturnsAnalysisBlock-class}
\alias{ImpliedFactorReturnsAnalysisBlock-class}
\title{Analysis Module for visualisting market wide implied factor
returns}
\description{
Generates ggplot of implied factor returns.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualRiskModelHandler",
              "VirtualImpliedFactorReturnsDataHandler"
}

