% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_state.r
\docType{class}
\name{ImpliedFactorReturnsStateAnalysisBlock-class}
\alias{ImpliedFactorReturnsStateAnalysisBlock-class}
\title{Analysis Module for visualisting implied factor state}
\description{
Generates ggplot of implied factor state.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualRiskModelHandler",
              "VirtualImpliedFactorReturnsDataHandler"
}

