% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_hedge_exposures.r
\docType{class}
\name{IndexPortfolioCoreFactorHedgeExposuresAnalysisBlock-class}
\alias{IndexPortfolioCoreFactorHedgeExposuresAnalysisBlock-class}
\title{Analyser for Index Portfolio risk factors exposure}
\description{
Computation block class to pull data required for portfolio factor exposure.
}
\details{
Inherits from "PortfolioCoreFactorHedgeExposuresAnalysisBlock"
}

