% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_exposures.r
\docType{class}
\name{IndexPortfolioFactorExposuresAnalysisBlock-class}
\alias{IndexPortfolioFactorExposuresAnalysisBlock-class}
\title{Analyser for Index Portfolio risk factors exposure}
\description{
Computation block class to pull data required for portfolio factor exposure.
}
\details{
Inherits from "PortfolioFactorExposuresAnalysisBlock"
}

