% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_returns.r
\docType{class}
\name{IndexPortfolioFactorReturnsAnalysisBlock-class}
\alias{IndexPortfolioFactorReturnsAnalysisBlock-class}
\title{Analysis Module for computation of Index Portfolio Returns Decomposition}
\description{
Computation block class to pull data required for portfolio
returns decomposition. Pulls data required for computation
and adds required columns. Generates ggplot with returns
decomposition to factors.
}
\details{
Inherits from "PortfolioFactorReturnsAnalysisBlock"
}

