% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk.r
\docType{class}
\name{IndexPortfolioInstrumentMCTRAnalysisBlock-class}
\alias{IndexPortfolioInstrumentMCTRAnalysisBlock-class}
\title{Analysis Module for computation of Index Portfolio Decomposition}
\description{
Computation block class to pull data required for index portfolio
variance decomposition. Pulls data required for computation
and adds required columns.
}
\details{
Inherits from "PortfolioInstrumentMCTRAnalysisBlock"
}

