% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{class}
\name{IndexPortfolioVarianceDecompositionAnalysisBlock-class}
\alias{IndexPortfolioVarianceDecompositionAnalysisBlock-class}
\title{Analysis Module for computation of Index Portfolio Decomposition}
\description{
Computation block class to pull data required for index portfolio
variance decomposition. Pulls data required for computation
and adds required columns.
}
\details{
Inherits from "PortfolioVarianceDecompositionAnalysisBlock"
}

