% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{methods}
\name{initialize,PortfolioVarianceFactorDecompositionData-method}
\alias{initialize,PortfolioVarianceFactorDecompositionData-method}
\title{Initialize method for "PortfolioVarianceFactorDecompositionData" class
copies risk model values to respective objects.}
\usage{
\S4method{initialize}{PortfolioVarianceFactorDecompositionData}(.Object)
}
\arguments{
\item{.Object, }{object of class "PortfolioVarianceFactorDecompositionData"}
}
\value{
\code{.Object} object of class "PortfolioVarianceFactorDecompositionData"
}
\description{
Initialize method for "PortfolioVarianceFactorDecompositionData" class
copies risk model values to respective objects.
}

