% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_return.r
\docType{class}
\name{MarketReturnAnalysisBlock-class}
\alias{MarketReturnAnalysisBlock-class}
\title{Analysis Module for computation of
market return of extended trades positions}
\description{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler",
              "VirtualMarketDataHandler"
}

