% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_timeseries.r
\docType{class}
\name{MarketStyleAnalysisBlock-class}
\alias{MarketStyleAnalysisBlock-class}
\title{Analysis computing MarketStyle data and generating timeseries plot}
\description{
Market style is z-score of first eigenvector of factor covariance matrix.
}
\details{
Computation block class to pull data required data for Market Style

Inherits from "VirtualAnalysisBlock",
              "VirtualRiskModelHandler",
              "VirtualMarketStyleDataHandler"
}

