% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_factor_stat.r
\docType{class}
\name{MarketStyleFactorStatisticAnalysisBlock-class}
\alias{MarketStyleFactorStatisticAnalysisBlock-class}
\title{Analysis computing MarketStyle data}
\description{
Market style is z-score of first eigenvector of factor covariance matrix.
}
\details{
Computation block class to pull data required data for Market Style

Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler"
}

