% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\docType{methods}
\name{modifyGGPlotFormat}
\alias{modifyGGPlotFormat}
\alias{modifyGGPlotFormat,VirtualAggregateAnalysisBlock-method}
\title{Modify the plot format: i.e. add gg functions the plot stream}
\usage{
modifyGGPlotFormat(object)

\S4method{modifyGGPlotFormat}{VirtualAggregateAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "VirtualAggregateAnalysisBlock"}
}
\value{
\code{object} object of class "VirtualAggregateAnalysisBlock"
}
\description{
Modify the plot format: i.e. add gg functions the plot stream
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualAggregateAnalysisBlock}: Modify the plot format: i.e. add gg functions the plot stream
}}

