% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\docType{methods}
\name{modifyGGPlotObject}
\alias{modifyGGPlotObject}
\alias{modifyGGPlotObject,VirtualAggregateAnalysisBlock-method}
\title{Modify the plot object: i.e. reassign to fields internal to
the ggplot}
\usage{
modifyGGPlotObject(object)

\S4method{modifyGGPlotObject}{VirtualAggregateAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "VirtualAggregateAnalysisBlock"}
}
\value{
\code{object} object of class "VirtualAggregateAnalysisBlock"
}
\description{
Modify the plot object: i.e. reassign to fields internal to
the ggplot
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualAggregateAnalysisBlock}: Modify the plot object: i.e. reassign to fields internal to
the ggplot
}}

