% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\docType{methods}
\name{modifyUIOptions}
\alias{modifyUIOptions}
\alias{modifyUIOptions,VirtualAggregateAnalysisBlock-method}
\title{Modify the ui options list}
\usage{
modifyUIOptions(object)

\S4method{modifyUIOptions}{VirtualAggregateAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "VirtualAggregateAnalysisBlock"}
}
\value{
\code{object} object of class "VirtualAggregateAnalysisBlock"
}
\description{
Modify the ui options list
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualAggregateAnalysisBlock}: Modify the ui options list
}}

