% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_cumulative_pnl.r
\docType{class}
\name{OffsidePositionCumulativePnLData-class}
\alias{OffsidePositionCumulativePnLData-class}
\title{OffsidePositionCumulativePnLData Reference Data class.}
\description{
Concrete S4 class storing data of Offside Market Positions
cumulative PnL
Generated only by OffsidePositionsCumulativePnLAnalysisBlock module.
}
\details{
Inherits from "VirtualPositionData"
}

