% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions.r
\docType{class}
\name{OffsidePositionData-class}
\alias{OffsidePositionData-class}
\title{Offside Positions Reference Data class.}
\description{
Concrete S4 class storing data of Offside Market Positions.
Generated only by OffsidePositionsAnalysisBlock module.
}
\details{
Inherits from "VirtualPositionData"
}

