% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_pnl_vs_days.r
\docType{class}
\name{OffsidePositionGainAndCategoryData-class}
\alias{OffsidePositionGainAndCategoryData-class}
\title{OffsidePositionGainAndCategoryData Reference Data class.}
\description{
Concrete S4 class storing data of Offside Market Positions
gain and category data
Generated only by OffsidePositionsPnLVsDaysAnalysisBlock module.
}
\details{
Inherits from "OffsidePositionGainData"
}

