% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_gain_vs_days.r
\docType{class}
\name{OffsidePositionGainData-class}
\alias{OffsidePositionGainData-class}
\title{OffsidePositionGainData Reference Data class.}
\description{
Concrete S4 class storing data of Offside Market Positions
gain Data
Generated only by OffsidePositionsGainVsDaysAnalysisBlock module.
}
\details{
Inherits from "VirtualPositionData"
}

