% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_relative_bps_per_month.r
\docType{class}
\name{OffsidePositionRelativeBpsData-class}
\alias{OffsidePositionRelativeBpsData-class}
\title{OffsidePositionRelativeBpsData Reference Data class.}
\description{
Concrete S4 class storing data of Offside Market Positions
Relative Gain in terms of Base Position Points (BPS)
Generated only by OffsidePositionsBpsPerMonthAnalysisBlock module.
}
\details{
Inherits from "VirtualPositionData"
}

