% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions.r
\docType{class}
\name{OffsidePositionsAnalysisBlock-class}
\alias{OffsidePositionsAnalysisBlock-class}
\title{Analysis Module for computation of Market Offside Positions}
\description{
Computation block class to pull data required for Computation of offside positions
Pulls data required for computation and adds required columns.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualPositionDataHandler",
              "VirtualPriceDataHandler",
              "VirtualMarketDataHandler"
}

