% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_relative_bps_per_month.r
\docType{class}
\name{OffsidePositionsBpsPerMonthAnalysisBlock-class}
\alias{OffsidePositionsBpsPerMonthAnalysisBlock-class}
\title{Analysis Module for computation of Market Offside Positions
Relative Gain in terms of Base Position Points (BPS)}
\description{
Computation block class to pull data required for Computation of offside positions
Pulls data required for computation and adds required columns.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualPositionDataHandler"
}

