% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_cumulative_pnl.r
\docType{class}
\name{OffsidePositionsCumulativePnLAnalysisBlock-class}
\alias{OffsidePositionsCumulativePnLAnalysisBlock-class}
\title{Analysis Module for computation of Market Offside Positions PnL}
\description{
Computation block class to pull data required for Computation of offside positions
Pulls data required for computation and adds required columns.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualPositionDataHandler"
}

