% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_gain_vs_days.r
\docType{class}
\name{OffsidePositionsGainVsDaysAnalysisBlock-class}
\alias{OffsidePositionsGainVsDaysAnalysisBlock-class}
\title{Analysis Module for computation of Market Offside Positions Gain}
\description{
Computation block class to pull data required for Computation of
offside positions Gain vs Days
}
\details{
Pulls data required for computation and adds required columns.

Inherits from "VirtualAnalysisBlock",
              "VirtualPositionDataHandler"
}

