% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{data}
\name{portfolio_decomposition_commodity_factors}
\alias{portfolio_decomposition_commodity_factors}
\title{List of portfolio decomposition commodity factors}
\format{An object of class \code{character} of length 1.}
\usage{
portfolio_decomposition_commodity_factors
}
\description{
List of portfolio decomposition commodity factors
}
\keyword{datasets}

