% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{data}
\name{portfolio_decomposition_composition_factors}
\alias{portfolio_decomposition_composition_factors}
\title{List of all portfolio decomposition composed factors}
\format{An object of class \code{character} of length 5.}
\usage{
portfolio_decomposition_composition_factors
}
\description{
List of all portfolio decomposition composed factors
}
\keyword{datasets}

