% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{data}
\name{portfolio_decomposition_currency_factors}
\alias{portfolio_decomposition_currency_factors}
\title{List of portfolio decomposition currency factors}
\format{An object of class \code{character} of length 16.}
\usage{
portfolio_decomposition_currency_factors
}
\description{
List of portfolio decomposition currency factors
}
\keyword{datasets}

