% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{data}
\name{portfolio_decomposition_factor_groups}
\alias{portfolio_decomposition_factor_groups}
\title{List of all portfolio decomposition factor groups}
\format{An object of class \code{list} of length 5.}
\usage{
portfolio_decomposition_factor_groups
}
\description{
List of all portfolio decomposition factor groups
}
\keyword{datasets}

