% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{data}
\name{portfolio_decomposition_market_factors}
\alias{portfolio_decomposition_market_factors}
\title{List of portfolio decomposition market factors}
\format{An object of class \code{character} of length 10.}
\usage{
portfolio_decomposition_market_factors
}
\description{
List of portfolio decomposition market factors
}
\keyword{datasets}

