% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{data}
\name{portfolio_decomposition_sector_factors}
\alias{portfolio_decomposition_sector_factors}
\title{List of portfolio decomposition sector factors}
\format{An object of class \code{character} of length 18.}
\usage{
portfolio_decomposition_sector_factors
}
\description{
List of portfolio decomposition sector factors
}
\keyword{datasets}

