% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_exposures.r
\docType{class}
\name{PortfolioCoreFactorExposuresData-class}
\alias{PortfolioCoreFactorExposuresData-class}
\title{PortfolioCoreFactorExposuresData Reference Data class.}
\description{
Concrete S4 class storing data of Portfolio factor exposures
Generated only by PortfolioCoreFactorExposuresAnalysisBlock module.
}
\details{
Inherits from "VirtualImpliedFactorReturnsData"
}

