% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_hedge_exposures.r
\docType{class}
\name{PortfolioCoreFactorHedgeExposuresData-class}
\alias{PortfolioCoreFactorHedgeExposuresData-class}
\title{PortfolioCoreFactorHedgeExposuresData Reference Data class.}
\description{
Concrete S4 class storing data of Portfolio factor exposures
Generated only by PortfolioCoreFactorHedgeExposuresAnalysisBlock module.
}
\details{
Inherits from "VirtualImpliedFactorReturnsData"
}

