% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_hedge_returns.r
\docType{class}
\name{PortfolioCoreFactorHedgeReturnsData-class}
\alias{PortfolioCoreFactorHedgeReturnsData-class}
\title{PortfolioCoreFactorHedgeReturnsData Reference Data class.}
\description{
Concrete S4 class storing data Portfolio Returns Decomposition
Data. Generated only within PortfolioCoreFactorHedgeReturnsAnalysisBlock
}
\details{
Inherits from "VirtualImpliedFactorReturnsData"
}

