% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_returns.r
\docType{class}
\name{PortfolioCoreFactorReturnsAnalysisBlock-class}
\alias{PortfolioCoreFactorReturnsAnalysisBlock-class}
\title{Analysis Module for computation of Portfolio Returns Decomposition}
\description{
Computation block class to pull data required for portfolio
returns decomposition. Pulls data required for computation
and adds required columns. Generates ggplot with returns
decomposition to factors.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualPortfolioDataHandler",
              "VirtualRiskModelHandler",
              "VirtualInstrumentBetasDataHandler",
              "VirtualImpliedFactorReturnsDataHandler"
}

