% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_returns.r
\docType{class}
\name{PortfolioCoreFactorReturnsData-class}
\alias{PortfolioCoreFactorReturnsData-class}
\title{PortfolioCoreFactorReturnsData Reference Data class.}
\description{
Concrete S4 class storing data Portfolio Returns Decomposition
Data. Generated only within PortfolioCoreFactorReturnsAnalysisBlock
}
\details{
Inherits from "VirtualImpliedFactorReturnsData"
}

