% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_exposure.r
\docType{class}
\name{PortfolioFactorExposureAnalysisBlock-class}
\alias{PortfolioFactorExposureAnalysisBlock-class}
\title{Analysis Module for extraction Factor Exposure of Portfolio}
\description{
Computation block class to pull data required for Computation Portfolio Factor
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualPortfolioDataHandler",
              "VirtualFactorExposureDataHandler"
}

