% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_exposures.r
\docType{class}
\name{PortfolioFactorExposuresData-class}
\alias{PortfolioFactorExposuresData-class}
\title{PortfolioFactorExposuresData Reference Data class.}
\description{
Concrete S4 class storing data of Portfolio factor exposures
Generated only by PortfolioFactorExposuresAnalysisBlock module.
}
\details{
Inherits from "VirtualImpliedFactorReturnsData"
}

