% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_returns.r
\docType{class}
\name{PortfolioFactorReturnsData-class}
\alias{PortfolioFactorReturnsData-class}
\title{PortfolioFactorReturnsData Reference Data class.}
\description{
Concrete S4 class storing data Portfolio Returns Decomposition
Data. Generated only within PortfolioFactorReturnsAnalysisBlock
}
\details{
Inherits from "VirtualImpliedFactorReturnsData"
}

