% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_state_prob_cond_loss.r
\docType{class}
\name{PortfolioFactorStateProbCondLossAnalysisBlock-class}
\alias{PortfolioFactorStateProbCondLossAnalysisBlock-class}
\title{Analysis Module for visualisting and computing
implied factor state probability conditioned on loss of
portfolio}
\description{
Generates ggplot of implied factor state.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualRiskModelHandler",
              "VirtualImpliedFactorReturnsDataHandler"
}

