% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r
\docType{class}
\name{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock-class}
\alias{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock-class}
\title{Analysis Module for computation of Portfolio Decomposition}
\description{
Computation block class to pull data required for portfolio
variance decomposition. Pulls data required for computation
and adds required columns.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualPortfolioDataHandler",
              "VirtualRiskModelHandler",
              "VirtualInstrumentBetasDataHandler",
              "VirtualFactorCorrelationDataHandler",
              "VirtualFactorVarianceDataHandler"
}

