% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r
\docType{class}
\name{PortfolioHedgeInstrumentCoreFactorsMCTRData-class}
\alias{PortfolioHedgeInstrumentCoreFactorsMCTRData-class}
\title{PortfolioVarianceFactorDecompositionData Reference Data class.}
\description{
Concrete S4 class storing data Portfolio Variance Decomposition
Data. Generated only within PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock
}
\details{
Inherits from "VirtualFactorVarianceData"
}

