% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_state_prob_cond_loss.r
\docType{class}
\name{PortfolioImpliedFactorReturnsStateProbData-class}
\alias{PortfolioImpliedFactorReturnsStateProbData-class}
\title{PortfolioImpliedFactorReturnsStateProbData Reference Data class.}
\description{
Concrete S4 class storing data Factor state probability
conditioned on
}
\details{
Inherits from "VirtualFactorVarianceData"
}

