% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{class}
\name{PortfolioInstrumentCoreFactorsMCTRData-class}
\alias{PortfolioInstrumentCoreFactorsMCTRData-class}
\title{PortfolioVarianceFactorDecompositionData Reference Data class.}
\description{
Concrete S4 class storing data Portfolio Variance Decomposition
Data. Generated only within PortfolioInstrumentCoreFactorsMCTRAnalysisBlock
}
\details{
Inherits from "VirtualFactorVarianceData"
}

