% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{class}
\name{PortfolioVarianceFactorDecompositionData-class}
\alias{PortfolioVarianceFactorDecompositionData-class}
\title{PortfolioVarianceFactorDecompositionData Reference Data class.}
\description{
Concrete S4 class storing data Portfolio Variance Decomposition
Data. Generated only within PortfolioVarianceDecompositionAnalysisBlock
}
\details{
Inherits from "VirtualFactorVarianceData"
}

