% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_revisits.r
\docType{class}
\name{PositionRevisitsAnalysisBlock-class}
\alias{PositionRevisitsAnalysisBlock-class}
\title{Analysis Module for computation of stock position revisits}
\description{
Computation block class that computes position revisits data.
Gets required data and generates ggplot with PnL per position
revisit.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler"
}

