% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_revisits_delta_prev_quarter.r
\docType{class}
\name{PositionRevisitsDeltaPrevQuarterAnalysisBlock-class}
\alias{PositionRevisitsDeltaPrevQuarterAnalysisBlock-class}
\title{Analysis Module for computation of stock position revisits change in quarter.}
\description{
Computation block class that computes position revisits data.
Computes change of stock revisits vs previous quarter.
Gets required data and generates ggplot with PnL per position
revisit and relative change of revisit trades.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler"
}

