% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_capital_distribution.r
\docType{class}
\name{PositionsHoldingCapitalDistributionAnalysisBlock-class}
\alias{PositionsHoldingCapitalDistributionAnalysisBlock-class}
\title{Analysis Module for computation of capital distribution over time.}
\description{
Computation block class that computes capital age curve.
Pools all required data and generates gg plot with capital age curve.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualPositionDataHandler"
}

