% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_day_zero_pnl.r
\docType{class}
\name{PositionsHoldingDayZeroPnLAnalysisBlock-class}
\alias{PositionsHoldingDayZeroPnLAnalysisBlock-class}
\title{Analysis Module for Computation position PnL at day zero of trade.}
\description{
Pools all required data and generates gg plot PnL at day zero.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualPositionDataHandler",
              "VirtualTradeDataHandler"
}

