% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_period.r
\docType{class}
\name{PositionsHoldingPeriodAnalysisBlock-class}
\alias{PositionsHoldingPeriodAnalysisBlock-class}
\title{Analysis Module for Computation position position holding period.}
\description{
Pools all required data and generates gg plot.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualPositionDataHandler"
}

