% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\name{.setOutputFrontendData}
\alias{.setOutputFrontendData}
\title{Set internaly data for ggplot object}
\usage{
.setOutputFrontendData(object, frontend_data)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}

\item{frontend_data}{object of class "data.frame"}
}
\value{
\code{object} object object of class "VirtualAnalysisBlock"
}
\description{
Private method to set frontend_data slot with data.frame used to provide to frontend services object
}

