% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\name{.setOutputGGPlot}
\alias{.setOutputGGPlot}
\title{Set internaly output ggplot object}
\usage{
.setOutputGGPlot(object, ggplot)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}

\item{ggplot}{object of class "ggplot"}
}
\value{
\code{object} object object of class "VirtualAnalysisBlock"
}
\description{
Private method to set output ggplot object slot
}

