% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\name{.setOutputGGPlotData}
\alias{.setOutputGGPlotData}
\title{Set internaly data for ggplot object}
\usage{
.setOutputGGPlotData(object, ggplot_data)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}

\item{ggplot_data}{object of class "data.frame"}
}
\value{
object object of class "VirtualAnalysisBlock"
}
\description{
Private method to set slot with data used to generate ggplot object.
}

