% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\name{.setOutputObject}
\alias{.setOutputObject}
\title{Set internaly output object}
\usage{
.setOutputObject(object, ref_data_object)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}

\item{ref_data_object}{object of class derived from "VirtualReferenceData"}
}
\value{
\code{object} object object of class "VirtualAnalysisBlock"
}
\description{
Private method to set output object slot
}

