% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades.r
\docType{methods}
\name{Process,AverageDownTradesAnalysisBlock-method}
\alias{Process,AverageDownTradesAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{AverageDownTradesAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "AverageDownTradesAnalysisBlock"}
}
\value{
\code{object} object object of class "AverageDownTradesAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

