% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades_focus.r
\docType{methods}
\name{Process,AverageDownTradesFocusAnalysisBlock-method}
\alias{Process,AverageDownTradesFocusAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{AverageDownTradesFocusAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "OffsidePositionsBpsPerMonthAnalysisBlock"}
}
\value{
\code{object} object object of class "OffsidePositionsBpsPerMonthAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

