% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buys_and_sells.r
\docType{methods}
\name{Process,BuysAndSellsAnalysisBlock-method}
\alias{Process,BuysAndSellsAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{BuysAndSellsAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "BuysAndSellsAnalysisBlock"}
}
\value{
\code{object} object object of class "BuysAndSellsAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

