% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trades.r
\docType{methods}
\name{Process,ExtendedTradesAnalysisBlock-method}
\alias{Process,ExtendedTradesAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{ExtendedTradesAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "ExtendedTradesAnalysisBlock"}
}
\value{
\code{object} object object of class "ExtendedTradesAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

