% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trades_summary.r
\docType{methods}
\name{Process,ExtendedTradesSummaryAnalysisBlock-method}
\alias{Process,ExtendedTradesSummaryAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{ExtendedTradesSummaryAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "ExtendedTradesSummaryAnalysisBlock"}
}
\value{
\code{object} object object of class "ExtendedTradesSummaryAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

