% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_state.r
\docType{methods}
\name{Process,ImpliedFactorReturnsStateAnalysisBlock-method}
\alias{Process,ImpliedFactorReturnsStateAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{ImpliedFactorReturnsStateAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "ImpliedFactorReturnsAnalysisBlock"}
}
\value{
\code{object} object object of class "ImpliedFactorReturnsAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

