% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_return.r
\docType{methods}
\name{Process,MarketReturnAnalysisBlock-method}
\alias{Process,MarketReturnAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{MarketReturnAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "MarketReturnAnalysisBlock"}
}
\value{
\code{object} object object of class "MarketReturnAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

