% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_timeseries.r
\docType{methods}
\name{Process,MarketStyleAnalysisBlock-method}
\alias{Process,MarketStyleAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{MarketStyleAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "MarketStyleAnalysisBlock"}
}
\value{
\code{object} object object of class "MarketStyleAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

