% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_factor_stat.r
\docType{methods}
\name{Process,MarketStyleFactorStatisticAnalysisBlock-method}
\alias{Process,MarketStyleFactorStatisticAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{MarketStyleFactorStatisticAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "MarketStyleFactorStatisticAnalysisBlock"}
}
\value{
\code{object} object object of class "MarketStyleFactorStatisticAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

