% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions.r
\docType{methods}
\name{Process,OffsidePositionsAnalysisBlock-method}
\alias{Process,OffsidePositionsAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{OffsidePositionsAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "OffsidePositionsAnalysisBlock"}
}
\value{
\code{object} object object of class "OffsidePositionsAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

