% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_relative_bps_per_month.r
\docType{methods}
\name{Process,OffsidePositionsBpsPerMonthAnalysisBlock-method}
\alias{Process,OffsidePositionsBpsPerMonthAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{OffsidePositionsBpsPerMonthAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "OffsidePositionsBpsPerMonthAnalysisBlock"}
}
\value{
\code{object} object object of class "OffsidePositionsBpsPerMonthAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

