% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_cumulative_pnl.r
\docType{methods}
\name{Process,OffsidePositionsCumulativePnLAnalysisBlock-method}
\alias{Process,OffsidePositionsCumulativePnLAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{OffsidePositionsCumulativePnLAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "OffsidePositionsCumulativePnLAnalysisBlock"}
}
\value{
\code{object} object object of class "OffsidePositionsCumulativePnLAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

