% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_gain_vs_days.r
\docType{methods}
\name{Process,OffsidePositionsGainVsDaysAnalysisBlock-method}
\alias{Process,OffsidePositionsGainVsDaysAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{OffsidePositionsGainVsDaysAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "OffsidePositionsGainVsDaysAnalysisBlock"}
}
\value{
\code{object} object object of class "OffsidePositionsGainVsDaysAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

