% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_pnl_vs_days.r
\docType{methods}
\name{Process,OffsidePositionsPnLVsDaysAnalysisBlock-method}
\alias{Process,OffsidePositionsPnLVsDaysAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{OffsidePositionsPnLVsDaysAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "OffsidePositionsPnLVsDaysAnalysisBlock"}
}
\value{
\code{object} object object of class "OffsidePositionsPnLVsDaysAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

