% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl_traded_in_long_short_hedge.r
\docType{methods}
\name{Process,PnLTradedInLongShortHedgeAnalysisBlock-method}
\alias{Process,PnLTradedInLongShortHedgeAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PnLTradedInLongShortHedgeAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PnLTradedInLongShortHedgeAnalysisBlock"}
}
\value{
\code{object} object object of class "PnLTradedInLongShortHedgeAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

