% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_exposures.r
\docType{methods}
\name{Process,PortfolioCoreFactorExposuresAnalysisBlock-method}
\alias{Process,PortfolioCoreFactorExposuresAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PortfolioCoreFactorExposuresAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PortfolioCoreFactorExposuresAnalysisBlock"}
}
\value{
\code{object} object object of class "PortfolioCoreFactorExposuresAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

