% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_hedge_exposures.r
\docType{methods}
\name{Process,PortfolioCoreFactorHedgeExposuresAnalysisBlock-method}
\alias{Process,PortfolioCoreFactorHedgeExposuresAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PortfolioCoreFactorHedgeExposuresAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PortfolioCoreFactorHedgeExposuresAnalysisBlock"}
}
\value{
\code{object} object object of class "PortfolioCoreFactorHedgeExposuresAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

