% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_hedge_returns.r
\docType{methods}
\name{Process,PortfolioCoreFactorHedgeReturnsAnalysisBlock-method}
\alias{Process,PortfolioCoreFactorHedgeReturnsAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PortfolioCoreFactorHedgeReturnsAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PortfolioCoreFactorHedgeReturnsAnalysisBlock"}
}
\value{
\code{object} object object of class "PortfolioCoreFactorHedgeReturnsAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

