% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_returns.r
\docType{methods}
\name{Process,PortfolioCoreFactorReturnsAnalysisBlock-method}
\alias{Process,PortfolioCoreFactorReturnsAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PortfolioCoreFactorReturnsAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PortfolioCoreFactorReturnsAnalysisBlock"}
}
\value{
\code{object} object object of class "PortfolioCoreFactorReturnsAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

