% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_exposure.r
\docType{methods}
\name{Process,PortfolioFactorExposureAnalysisBlock-method}
\alias{Process,PortfolioFactorExposureAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PortfolioFactorExposureAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PortfolioFactorExposureAnalysisBlock"}
}
\value{
\code{object} object object of class "PortfolioFactorExposureAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

