% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_exposures.r
\docType{methods}
\name{Process,PortfolioFactorExposuresAnalysisBlock-method}
\alias{Process,PortfolioFactorExposuresAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PortfolioFactorExposuresAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PortfolioFactorExposuresAnalysisBlock"}
}
\value{
\code{object} object object of class "PortfolioFactorExposuresAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

