% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_returns.r
\docType{methods}
\name{Process,PortfolioFactorReturnsAnalysisBlock-method}
\alias{Process,PortfolioFactorReturnsAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PortfolioFactorReturnsAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PortfolioFactorReturnsAnalysisBlock"}
}
\value{
\code{object} object object of class "PortfolioFactorReturnsAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

