% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{Process,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock-method}
\alias{Process,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{

  \S4method{Process}{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock"}
}
\value{
\code{object} object object of class "PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

