% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{Process,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock-method}
\alias{Process,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PortfolioInstrumentCoreFactorsMCTRAnalysisBlock"}
}
\value{
\code{object} object object of class "PortfolioInstrumentCoreFactorsMCTRAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

