% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk.r
\docType{methods}
\name{Process,PortfolioInstrumentMCTRAnalysisBlock-method}
\alias{Process,PortfolioInstrumentMCTRAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PortfolioInstrumentMCTRAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PortfolioInstrumentMCTRAnalysisBlock"}
}
\value{
\code{object} object object of class "PortfolioInstrumentMCTRAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

