% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{methods}
\name{Process,PortfolioVarianceDecompositionAnalysisBlock-method}
\alias{Process,PortfolioVarianceDecompositionAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PortfolioVarianceDecompositionAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PortfolioVarianceDecompositionAnalysisBlock"}
}
\value{
\code{object} object object of class "PortfolioVarianceDecompositionAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

