% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_revisits_delta_prev_quarter.r
\docType{methods}
\name{Process,PositionRevisitsDeltaPrevQuarterAnalysisBlock-method}
\alias{Process,PositionRevisitsDeltaPrevQuarterAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PositionRevisitsDeltaPrevQuarterAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PositionRevisitsDeltaPrevQuarterAnalysisBlock"}
}
\value{
\code{object} object object of class "PositionRevisitsDeltaPrevQuarterAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

