% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_capital_distribution.r
\docType{methods}
\name{Process,PositionsHoldingCapitalDistributionAnalysisBlock-method}
\alias{Process,PositionsHoldingCapitalDistributionAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PositionsHoldingCapitalDistributionAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PositionsHoldingCapitalDistributionAnalysisBlock"}
}
\value{
\code{object} object object of class "PositionsHoldingCapitalDistributionAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

