% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_day_zero_pnl.r
\docType{methods}
\name{Process,PositionsHoldingDayZeroPnLAnalysisBlock-method}
\alias{Process,PositionsHoldingDayZeroPnLAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PositionsHoldingDayZeroPnLAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PositionsHoldingDayZeroPnLAnalysisBlock"}
}
\value{
\code{object} object object of class "PositionsHoldingDayZeroPnLAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

