% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_period.r
\docType{methods}
\name{Process,PositionsHoldingPeriodAnalysisBlock-method}
\alias{Process,PositionsHoldingPeriodAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PositionsHoldingPeriodAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "PositionsHoldingPeriodAnalysisBlock"}
}
\value{
\code{object} object object of class "PositionsHoldingPeriodAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

