% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_market_return.r
\docType{methods}
\name{Process,RelativeMarketReturnAnalysisBlock-method}
\alias{Process,RelativeMarketReturnAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{RelativeMarketReturnAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "RelativeMarketReturnAnalysisBlock"}
}
\value{
\code{object} object object of class "RelativeMarketReturnAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

