% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown.r
\docType{methods}
\name{Process,StrategyBreakdownAnalysisBlock-method}
\alias{Process,StrategyBreakdownAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{StrategyBreakdownAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "StrategyBreakdownAnalysisBlock"}
}
\value{
\code{object} object object of class "StrategyBreakdownAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

