% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_aum_and_turnover.r
\docType{methods}
\name{Process,StrategyBreakdownAUMAndTurnoverAnalysisBlock-method}
\alias{Process,StrategyBreakdownAUMAndTurnoverAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{StrategyBreakdownAUMAndTurnoverAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "StrategyBreakdownAUMAndTurnoverAnalysisBlock"}
}
\value{
\code{object} object object of class "StrategyBreakdownAUMAndTurnoverAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

