% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_pnl_and_turnover_by_event.r
\docType{methods}
\name{Process,StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock-method}
\alias{Process,StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{

  \S4method{Process}{StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock"}
}
\value{
\code{object} object object of class "StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

