% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_pnl_on_trade_day_per_signal.r
\docType{methods}
\name{Process,StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock-method}
\alias{Process,StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{

  \S4method{Process}{StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock"}
}
\value{
\code{object} object object of class "StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

