% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_total_pnl_and_pnl_delta.r
\docType{methods}
\name{Process,StrategyBreakdownTotalAndDeltaPnLAnalysisBlock-method}
\alias{Process,StrategyBreakdownTotalAndDeltaPnLAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{StrategyBreakdownTotalAndDeltaPnLAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "StrategyBreakdownTotalAndDeltaPnLAnalysisBlock"}
}
\value{
\code{object} object object of class "StrategyBreakdownTotalAndDeltaPnLAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

