% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_value_traded_per_signal.r
\docType{methods}
\name{Process,StrategyBreakdownValueTradedPerSignalAnalysisBlock-method}
\alias{Process,StrategyBreakdownValueTradedPerSignalAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{StrategyBreakdownValueTradedPerSignalAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "StrategyBreakdownValueTradedPerSignalAnalysisBlock"}
}
\value{
\code{object} object object of class "StrategyBreakdownValueTradedPerSignalAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

