% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades_performance_on_results_day.r
\docType{methods}
\name{Process,TradesPerformanceOnResultsDayAnalysisBlock-method}
\alias{Process,TradesPerformanceOnResultsDayAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{TradesPerformanceOnResultsDayAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "TradesPerformanceOnResultsDayAnalysisBlock"}
}
\value{
\code{object} object object of class "TradesPerformanceOnResultsDayAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

