% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades_pnl_outof_on_results_day.r
\docType{methods}
\name{Process,TradesPnLOutOfOnResultsDayAnalysisBlock-method}
\alias{Process,TradesPnLOutOfOnResultsDayAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{TradesPnLOutOfOnResultsDayAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "TradesPnLOutOfOnResultsDayAnalysisBlock"}
}
\value{
\code{object} object object of class "TradesPnLOutOfOnResultsDayAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

