% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_traded_in_long_short_hedge.r
\docType{methods}
\name{Process,ValueTradedInLongShortHedgeAnalysisBlock-method}
\alias{Process,ValueTradedInLongShortHedgeAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{ValueTradedInLongShortHedgeAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "ValueTradedInLongShortHedgeAnalysisBlock"}
}
\value{
\code{object} object object of class "ValueTradedInLongShortHedgeAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

