% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\docType{methods}
\name{Process,VirtualAggregateAnalysisBlock-method}
\alias{Process,VirtualAggregateAnalysisBlock-method}
\title{Trigger computation of analysis data: Overridden to have not function because
this does not make sense in the context of aggregate blocks.}
\usage{
\S4method{Process}{VirtualAggregateAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "VirtualAggregateAnalysisBlock"}
}
\value{
\code{object} object object of class "VirtualAggregateAnalysisBlock"
}
\description{
Trigger computation of analysis data: Overridden to have not function because
this does not make sense in the context of aggregate blocks.
}

