% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r, R/offside_positions.r, R/average_down_trades.r, R/average_down_trades_focus.r, R/extended_trades.r, R/buys_and_sells.r, R/extended_trades_summary.r, R/implied_factor_returns.r, R/market_return.r, R/portfolio_variance_decomposition.r, R/market_style_factor_stat.r, R/market_style_timeseries.r, R/offside_positions_gain_vs_days.r, R/offside_positions_cumulative_pnl.r, R/offside_positions_pnl_vs_days.r, R/offside_positions_relative_bps_per_month.r, R/value_traded_in_long_short_hedge.r, R/pnl_traded_in_long_short_hedge.r, R/portfolio_core_factor_exposures.r, R/portfolio_core_factor_hedge_exposures.r, R/portfolio_core_factor_hedge_returns.r, R/portfolio_core_factor_returns.r, R/portfolio_factor_exposure.r, R/portfolio_factor_exposures.r, R/portfolio_factor_returns.r, R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r, R/portfolio_instrument_marginal_contribution_to_risk.r, R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r, R/position_revisits.r, R/position_revisits_delta_prev_quarter.r, R/positions_holding_capital_distribution.r, R/positions_holding_day_zero_pnl.r, R/positions_holding_period.r, R/relative_market_return.r, R/strategy_breakdown.r, R/strategy_breakdown_aum_and_turnover.r, R/strategy_breakdown_value_traded_per_signal.r, R/strategy_breakdown_pnl_and_turnover_by_event.r, R/strategy_breakdown_pnl_on_trade_day_per_signal.r, R/strategy_breakdown_signal_characteristic_and_effeciveness.r, R/strategy_breakdown_total_pnl_and_pnl_delta.r, R/trades_performance_on_results_day.r, R/trades_pnl_outof_on_results_day.r
\docType{methods}
\name{Process}
\alias{Process}
\alias{Process,AverageDownTradesAnalysisBlock-method}
\alias{Process,AverageDownTradesFocusAnalysisBlock-method}
\alias{Process,BuysAndSellsAnalysisBlock-method}
\alias{Process,ExtendedTradesAnalysisBlock-method}
\alias{Process,ExtendedTradesSummaryAnalysisBlock-method}
\alias{Process,ImpliedFactorReturnsAnalysisBlock-method}
\alias{Process,MarketReturnAnalysisBlock-method}
\alias{Process,MarketStyleAnalysisBlock-method}
\alias{Process,MarketStyleFactorStatisticAnalysisBlock-method}
\alias{Process,OffsidePositionsAnalysisBlock-method}
\alias{Process,OffsidePositionsBpsPerMonthAnalysisBlock-method}
\alias{Process,OffsidePositionsCumulativePnLAnalysisBlock-method}
\alias{Process,OffsidePositionsGainVsDaysAnalysisBlock-method}
\alias{Process,OffsidePositionsPnLVsDaysAnalysisBlock-method}
\alias{Process,PnLTradedInLongShortHedgeAnalysisBlock-method}
\alias{Process,PortfolioCoreFactorExposuresAnalysisBlock-method}
\alias{Process,PortfolioCoreFactorHedgeExposuresAnalysisBlock-method}
\alias{Process,PortfolioCoreFactorHedgeReturnsAnalysisBlock-method}
\alias{Process,PortfolioCoreFactorReturnsAnalysisBlock-method}
\alias{Process,PortfolioFactorExposureAnalysisBlock-method}
\alias{Process,PortfolioFactorExposuresAnalysisBlock-method}
\alias{Process,PortfolioFactorReturnsAnalysisBlock-method}
\alias{Process,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock-method}
\alias{Process,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock-method}
\alias{Process,PortfolioInstrumentMCTRAnalysisBlock-method}
\alias{Process,PortfolioVarianceDecompositionAnalysisBlock-method}
\alias{Process,PositionRevisitsAnalysisBlock-method}
\alias{Process,PositionRevisitsDeltaPrevQuarterAnalysisBlock-method}
\alias{Process,PositionsHoldingCapitalDistributionAnalysisBlock-method}
\alias{Process,PositionsHoldingDayZeroPnLAnalysisBlock-method}
\alias{Process,PositionsHoldingPeriodAnalysisBlock-method}
\alias{Process,RelativeMarketReturnAnalysisBlock-method}
\alias{Process,StrategyBreakdownAUMAndTurnoverAnalysisBlock-method}
\alias{Process,StrategyBreakdownAnalysisBlock-method}
\alias{Process,StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock-method}
\alias{Process,StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock-method}
\alias{Process,StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock-method}
\alias{Process,StrategyBreakdownTotalAndDeltaPnLAnalysisBlock-method}
\alias{Process,StrategyBreakdownValueTradedPerSignalAnalysisBlock-method}
\alias{Process,TradesPerformanceOnResultsDayAnalysisBlock-method}
\alias{Process,TradesPnLOutOfOnResultsDayAnalysisBlock-method}
\alias{Process,ValueTradedInLongShortHedgeAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
Process(object)

\S4method{Process}{OffsidePositionsAnalysisBlock}(object)

\S4method{Process}{AverageDownTradesAnalysisBlock}(object)

\S4method{Process}{AverageDownTradesFocusAnalysisBlock}(object)

\S4method{Process}{ExtendedTradesAnalysisBlock}(object)

\S4method{Process}{BuysAndSellsAnalysisBlock}(object)

\S4method{Process}{ExtendedTradesSummaryAnalysisBlock}(object)

\S4method{Process}{ImpliedFactorReturnsAnalysisBlock}(object)

\S4method{Process}{MarketReturnAnalysisBlock}(object)

\S4method{Process}{PortfolioVarianceDecompositionAnalysisBlock}(object)

\S4method{Process}{MarketStyleFactorStatisticAnalysisBlock}(object)

\S4method{Process}{MarketStyleAnalysisBlock}(object)

\S4method{Process}{OffsidePositionsGainVsDaysAnalysisBlock}(object)

\S4method{Process}{OffsidePositionsCumulativePnLAnalysisBlock}(object)

\S4method{Process}{OffsidePositionsPnLVsDaysAnalysisBlock}(object)

\S4method{Process}{OffsidePositionsBpsPerMonthAnalysisBlock}(object)

\S4method{Process}{ValueTradedInLongShortHedgeAnalysisBlock}(object)

\S4method{Process}{PnLTradedInLongShortHedgeAnalysisBlock}(object)

\S4method{Process}{PortfolioCoreFactorExposuresAnalysisBlock}(object)

\S4method{Process}{PortfolioCoreFactorHedgeExposuresAnalysisBlock}(object)

\S4method{Process}{PortfolioCoreFactorHedgeReturnsAnalysisBlock}(object)

\S4method{Process}{PortfolioCoreFactorReturnsAnalysisBlock}(object)

\S4method{Process}{PortfolioFactorExposureAnalysisBlock}(object)

\S4method{Process}{PortfolioFactorExposuresAnalysisBlock}(object)

\S4method{Process}{PortfolioFactorReturnsAnalysisBlock}(object)


  \S4method{Process}{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock}(object)

\S4method{Process}{PortfolioInstrumentMCTRAnalysisBlock}(object)

\S4method{Process}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock}(object)

\S4method{Process}{PositionRevisitsAnalysisBlock}(object)

\S4method{Process}{PositionRevisitsDeltaPrevQuarterAnalysisBlock}(object)

\S4method{Process}{PositionsHoldingCapitalDistributionAnalysisBlock}(object)

\S4method{Process}{PositionsHoldingDayZeroPnLAnalysisBlock}(object)

\S4method{Process}{PositionsHoldingPeriodAnalysisBlock}(object)

\S4method{Process}{RelativeMarketReturnAnalysisBlock}(object)

\S4method{Process}{StrategyBreakdownAnalysisBlock}(object)

\S4method{Process}{StrategyBreakdownAUMAndTurnoverAnalysisBlock}(object)

\S4method{Process}{StrategyBreakdownValueTradedPerSignalAnalysisBlock}(object)


  \S4method{Process}{StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock}(object)


  \S4method{Process}{StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock}(object)


  \S4method{Process}{StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock}(object)

\S4method{Process}{StrategyBreakdownTotalAndDeltaPnLAnalysisBlock}(object)

\S4method{Process}{TradesPerformanceOnResultsDayAnalysisBlock}(object)

\S4method{Process}{TradesPnLOutOfOnResultsDayAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class derived from "VirtualAnalysisBlock"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}
\section{Methods (by class)}{
\itemize{
\item \code{OffsidePositionsAnalysisBlock}: Trigger computation of analysis data.

\item \code{AverageDownTradesAnalysisBlock}: Trigger computation of analysis data.

\item \code{AverageDownTradesFocusAnalysisBlock}: Trigger computation of analysis data.

\item \code{ExtendedTradesAnalysisBlock}: Trigger computation of analysis data.

\item \code{BuysAndSellsAnalysisBlock}: Trigger computation of analysis data.

\item \code{ExtendedTradesSummaryAnalysisBlock}: Trigger computation of analysis data.

\item \code{ImpliedFactorReturnsAnalysisBlock}: Trigger computation of analysis data.

\item \code{MarketReturnAnalysisBlock}: Trigger computation of analysis data.

\item \code{PortfolioVarianceDecompositionAnalysisBlock}: Trigger computation of analysis data.

\item \code{MarketStyleFactorStatisticAnalysisBlock}: Trigger computation of analysis data.

\item \code{MarketStyleAnalysisBlock}: Trigger computation of analysis data.

\item \code{OffsidePositionsGainVsDaysAnalysisBlock}: Trigger computation of analysis data.

\item \code{OffsidePositionsCumulativePnLAnalysisBlock}: Trigger computation of analysis data.

\item \code{OffsidePositionsPnLVsDaysAnalysisBlock}: Trigger computation of analysis data.

\item \code{OffsidePositionsBpsPerMonthAnalysisBlock}: Trigger computation of analysis data.

\item \code{ValueTradedInLongShortHedgeAnalysisBlock}: Trigger computation of analysis data.

\item \code{PnLTradedInLongShortHedgeAnalysisBlock}: Trigger computation of analysis data.

\item \code{PortfolioCoreFactorExposuresAnalysisBlock}: Trigger computation of analysis data.

\item \code{PortfolioCoreFactorHedgeExposuresAnalysisBlock}: Trigger computation of analysis data.

\item \code{PortfolioCoreFactorHedgeReturnsAnalysisBlock}: Trigger computation of analysis data.

\item \code{PortfolioCoreFactorReturnsAnalysisBlock}: Trigger computation of analysis data.

\item \code{PortfolioFactorExposureAnalysisBlock}: Trigger computation of analysis data.

\item \code{PortfolioFactorExposuresAnalysisBlock}: Trigger computation of analysis data.

\item \code{PortfolioFactorReturnsAnalysisBlock}: Trigger computation of analysis data.

\item \code{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock}: Trigger computation of analysis data.

\item \code{PortfolioInstrumentMCTRAnalysisBlock}: Trigger computation of analysis data.

\item \code{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock}: Trigger computation of analysis data.

\item \code{PositionRevisitsAnalysisBlock}: Trigger computation of analysis data.

\item \code{PositionRevisitsDeltaPrevQuarterAnalysisBlock}: Trigger computation of analysis data.

\item \code{PositionsHoldingCapitalDistributionAnalysisBlock}: Trigger computation of analysis data.

\item \code{PositionsHoldingDayZeroPnLAnalysisBlock}: Trigger computation of analysis data.

\item \code{PositionsHoldingPeriodAnalysisBlock}: Trigger computation of analysis data.

\item \code{RelativeMarketReturnAnalysisBlock}: Trigger computation of analysis data.

\item \code{StrategyBreakdownAnalysisBlock}: Trigger computation of analysis data.

\item \code{StrategyBreakdownAUMAndTurnoverAnalysisBlock}: Trigger computation of analysis data.

\item \code{StrategyBreakdownValueTradedPerSignalAnalysisBlock}: Trigger computation of analysis data.

\item \code{StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock}: Trigger computation of analysis data.

\item \code{StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock}: Trigger computation of analysis data.

\item \code{StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock}: Trigger computation of analysis data.

\item \code{StrategyBreakdownTotalAndDeltaPnLAnalysisBlock}: Trigger computation of analysis data.

\item \code{TradesPerformanceOnResultsDayAnalysisBlock}: Trigger computation of analysis data.

\item \code{TradesPnLOutOfOnResultsDayAnalysisBlock}: Trigger computation of analysis data.
}}

