% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_market_return.r
\docType{class}
\name{RelativeMarketReturnAnalysisBlock-class}
\alias{RelativeMarketReturnAnalysisBlock-class}
\title{Analysis Module for computation of relative
market return of extended trades positions}
\description{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler",
              "VirtualExtendedTradeDataHandler",
              "VirtualMarketDataHandler"
}

