% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_value_traded_per_signal.r
\docType{methods}
\name{setEventDataObject,StrategyBreakdownValueTradedPerSignalAnalysisBlock,EventData-method}
\alias{setEventDataObject,StrategyBreakdownValueTradedPerSignalAnalysisBlock,EventData-method}
\title{Set event_data object in object slot}
\usage{

  \S4method{setEventDataObject}{StrategyBreakdownValueTradedPerSignalAnalysisBlock,EventData}(object,
  event_data)
}
\arguments{
\item{object}{object of class "StrategyBreakdownValueTradedPerSignalAnalysisBlock"}

\item{event_data}{object of class "EventData"}
}
\value{
\code{object} object of class "StrategyBreakdownValueTradedPerSignalAnalysisBlock"
}
\description{
Public method to set event_data slot with "EventData"
class object
}

