% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades_performance_on_results_day.r
\docType{methods}
\name{setEventDataObject,TradesPerformanceOnResultsDayAnalysisBlock,EventData-method}
\alias{setEventDataObject,TradesPerformanceOnResultsDayAnalysisBlock,EventData-method}
\title{Set event_data object in object slot}
\usage{

  \S4method{setEventDataObject}{TradesPerformanceOnResultsDayAnalysisBlock,EventData}(object,
  event_data)
}
\arguments{
\item{object}{object of class "TradesPerformanceOnResultsDayAnalysisBlock"}

\item{event_data}{object of class "EventData"}
}
\value{
\code{object} object of class "TradesPerformanceOnResultsDayAnalysisBlock"
}
\description{
Public method to set event_data slot with "EventData"
class object
}

