% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/strategy_breakdown_value_traded_per_signal.r, R/trades_performance_on_results_day.r
\docType{methods}
\name{setEventDataObject}
\alias{setEventDataObject}
\alias{setEventDataObject,StrategyBreakdownValueTradedPerSignalAnalysisBlock,EventData-method}
\alias{setEventDataObject,TradesPerformanceOnResultsDayAnalysisBlock,EventData-method}
\alias{setEventDataObject,VirtualAnalysisBlock,EventData-method}
\title{Set event_data object in object slot}
\usage{

  \S4method{setEventDataObject}{StrategyBreakdownValueTradedPerSignalAnalysisBlock,EventData}(object,
  event_data)


  \S4method{setEventDataObject}{TradesPerformanceOnResultsDayAnalysisBlock,EventData}(object,
  event_data)
}
\arguments{
\item{object}{object of class derived from"VirtualAnalysisBlock"}

\item{event_data}{object of class derived from "EventData"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"

\code{object} object of class "TradesPerformanceOnResultsDayAnalysisBlock"
}
\description{
Public method to set event_data slot with object of class
derived from "EventData"
}
\section{Methods (by class)}{
\itemize{
\item \code{object = StrategyBreakdownValueTradedPerSignalAnalysisBlock,event_data = EventData}: Set event_data object in object slot

\item \code{object = TradesPerformanceOnResultsDayAnalysisBlock,event_data = EventData}: Set event_data object in object slot
}}

