% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trades_summary.r
\docType{methods}
\name{setExtendedTradeDataObject,ExtendedTradesSummaryAnalysisBlock,ExtendedTradeData-method}
\alias{setExtendedTradeDataObject,ExtendedTradesSummaryAnalysisBlock,ExtendedTradeData-method}
\title{Set ex_trade_data object in object slot}
\usage{

  \S4method{setExtendedTradeDataObject}{ExtendedTradesSummaryAnalysisBlock,ExtendedTradeData}(object,
  ex_trade_data)
}
\arguments{
\item{object}{object of class "ExtendedTradesSummaryAnalysisBlock"}

\item{ex_trade_data}{object of class "ExtendedTradeData"}
}
\value{
\code{object} object of class "ExtendedTradesSummaryAnalysisBlock"
}
\description{
Public method to set ex_trade_data slot with "ExtendedTradeData"
class object
}

