% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_market_return.r
\docType{methods}
\name{setExtendedTradeDataObject,RelativeMarketReturnAnalysisBlock,TradesExtendedReturnPerMonth-method}
\alias{setExtendedTradeDataObject,RelativeMarketReturnAnalysisBlock,TradesExtendedReturnPerMonth-method}
\title{Set ex_trade_data object in object slot}
\usage{

  \S4method{setExtendedTradeDataObject}{RelativeMarketReturnAnalysisBlock,TradesExtendedReturnPerMonth}(object,
  ex_trade_data)
}
\arguments{
\item{object}{object of class "RelativeMarketReturnAnalysisBlock"}

\item{ex_trade_data}{object of class "TradesExtendedReturnPerMonth"}
}
\value{
\code{object} object of class "RelativeMarketReturnAnalysisBlock"
}
\description{
Public method to set ex_trade_data slot with "TradesExtendedReturnPerMonth"
class object
}

