% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/extended_trades_summary.r, R/relative_market_return.r
\docType{methods}
\name{setExtendedTradeDataObject}
\alias{setExtendedTradeDataObject}
\alias{setExtendedTradeDataObject,ExtendedTradesSummaryAnalysisBlock,ExtendedTradeData-method}
\alias{setExtendedTradeDataObject,RelativeMarketReturnAnalysisBlock,TradesExtendedReturnPerMonth-method}
\alias{setExtendedTradeDataObject,VirtualAnalysisBlock,ExtendedTradeData-method}
\title{Set ex_trade_data object in object slot}
\usage{

  \S4method{setExtendedTradeDataObject}{ExtendedTradesSummaryAnalysisBlock,ExtendedTradeData}(object,
  ex_trade_data)


  \S4method{setExtendedTradeDataObject}{RelativeMarketReturnAnalysisBlock,TradesExtendedReturnPerMonth}(object,
  ex_trade_data)
}
\arguments{
\item{object}{object of class derived from"VirtualAnalysisBlock"}

\item{ex_trade_data}{object of class "ExtendedTradeData"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"
}
\description{
Public method to set ex_trade_data slot with "ExtendedTradeData"
class or derived class object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = ExtendedTradesSummaryAnalysisBlock,ex_trade_data = ExtendedTradeData}: Set ex_trade_data object in object slot

\item \code{object = RelativeMarketReturnAnalysisBlock,ex_trade_data = TradesExtendedReturnPerMonth}: Set ex_trade_data object in object slot
}}

