% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setFactorCorrelationDataObject,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,FactorCorrelationData-method}
\alias{setFactorCorrelationDataObject,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,FactorCorrelationData-method}
\title{Set factor_correlation object in object slot}
\usage{

  \S4method{setFactorCorrelationDataObject}{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,FactorCorrelationData}(object,
  factor_correlation)
}
\arguments{
\item{object}{object of class "PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock"}

\item{factor_correlation}{object of class "FactorCorrelationData"}
}
\value{
\code{object} object of class "PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock"
}
\description{
Public method to set factor_correlation slot with "InstrumentBetasData"
class object
}

