% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setFactorCorrelationDataObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,FactorCorrelationData-method}
\alias{setFactorCorrelationDataObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,FactorCorrelationData-method}
\title{Set factor_correlation object in object slot}
\usage{

  \S4method{setFactorCorrelationDataObject}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,FactorCorrelationData}(object,
  factor_correlation)
}
\arguments{
\item{object}{object of class "PortfolioInstrumentCoreFactorsMCTRAnalysisBlock"}

\item{factor_correlation}{object of class "FactorCorrelationData"}
}
\value{
\code{object} object of class "PortfolioInstrumentCoreFactorsMCTRAnalysisBlock"
}
\description{
Public method to set factor_correlation slot with "InstrumentBetasData"
class object
}

