% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r, R/portfolio_instrument_marginal_contribution_to_risk.r, R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setFactorCorrelationDataObject}
\alias{setFactorCorrelationDataObject}
\alias{setFactorCorrelationDataObject,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,FactorCorrelationData-method}
\alias{setFactorCorrelationDataObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,FactorCorrelationData-method}
\alias{setFactorCorrelationDataObject,PortfolioInstrumentMCTRAnalysisBlock,FactorCorrelationData-method}
\alias{setFactorCorrelationDataObject,VirtualAnalysisBlock,FactorCorrelationData-method}
\title{Set factor_correlation object in object slot}
\usage{

  \S4method{setFactorCorrelationDataObject}{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,FactorCorrelationData}(object,
  factor_correlation)


  \S4method{setFactorCorrelationDataObject}{PortfolioInstrumentMCTRAnalysisBlock,FactorCorrelationData}(object,
  factor_correlation)


  \S4method{setFactorCorrelationDataObject}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,FactorCorrelationData}(object,
  factor_correlation)
}
\arguments{
\item{object}{object of class derived from"VirtualAnalysisBlock"}

\item{factor_correlation}{object of class "FactorCorrelationData"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"
}
\description{
Public method to set factor_correlation slot with "FactorCorrelationData"
class object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,factor_correlation = FactorCorrelationData}: Set factor correlation data in object slot

\item \code{object = PortfolioInstrumentMCTRAnalysisBlock,factor_correlation = FactorCorrelationData}: Set factor correlation data in object slot

\item \code{object = PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,factor_correlation = FactorCorrelationData}: Set factor correlation data in object slot
}}

