% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setFactorVarianceDataObject,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,FactorVarianceData-method}
\alias{setFactorVarianceDataObject,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,FactorVarianceData-method}
\title{Set factor_variance object in object slot}
\usage{

  \S4method{setFactorVarianceDataObject}{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,FactorVarianceData}(object,
  factor_variance)
}
\arguments{
\item{object}{object of class "PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock"}

\item{factor_variance}{object of class "FactorVarianceData"}
}
\value{
\code{object} object of class "PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock"
}
\description{
Public method to set factor_variance slot with "InstrumentBetasData"
class object
}

