% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk.r
\docType{methods}
\name{setFactorVarianceDataObject,PortfolioInstrumentMCTRAnalysisBlock,FactorVarianceData-method}
\alias{setFactorVarianceDataObject,PortfolioInstrumentMCTRAnalysisBlock,FactorVarianceData-method}
\title{Set factor_variance object in object slot}
\usage{

  \S4method{setFactorVarianceDataObject}{PortfolioInstrumentMCTRAnalysisBlock,FactorVarianceData}(object,
  factor_variance)
}
\arguments{
\item{object}{object of class "PortfolioInstrumentMCTRAnalysisBlock"}

\item{factor_variance}{object of class "FactorVarianceData"}
}
\value{
\code{object} object of class "PortfolioInstrumentMCTRAnalysisBlock"
}
\description{
Public method to set factor_variance slot with "InstrumentBetasData"
class object
}

