% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r, R/portfolio_instrument_marginal_contribution_to_risk.r, R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setFactorVarianceDataObject}
\alias{setFactorVarianceDataObject}
\alias{setFactorVarianceDataObject,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,FactorVarianceData-method}
\alias{setFactorVarianceDataObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,FactorVarianceData-method}
\alias{setFactorVarianceDataObject,PortfolioInstrumentMCTRAnalysisBlock,FactorVarianceData-method}
\alias{setFactorVarianceDataObject,VirtualAnalysisBlock,FactorVarianceData-method}
\title{Set factor_variance object in object slot}
\usage{

  \S4method{setFactorVarianceDataObject}{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,FactorVarianceData}(object,
  factor_variance)


  \S4method{setFactorVarianceDataObject}{PortfolioInstrumentMCTRAnalysisBlock,FactorVarianceData}(object,
  factor_variance)


  \S4method{setFactorVarianceDataObject}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,FactorVarianceData}(object,
  factor_variance)
}
\arguments{
\item{object}{object of class derived from"VirtualAnalysisBlock"}

\item{factor_variance}{object of class "FactorVarianceData"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"
}
\description{
Public method to set factor_variance slot with "FactorVarianceData"
class object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,factor_variance = FactorVarianceData}: Set factor variance data in object slot

\item \code{object = PortfolioInstrumentMCTRAnalysisBlock,factor_variance = FactorVarianceData}: Set factor variance data in object slot

\item \code{object = PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,factor_variance = FactorVarianceData}: Set factor variance data in object slot
}}

