% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/portfolio_factor_state_prob_cond_loss.r
\docType{methods}
\name{setImpliedFactorReturnsDataObject}
\alias{setImpliedFactorReturnsDataObject}
\alias{setImpliedFactorReturnsDataObject,PortfolioFactorStateProbCondLossAnalysisBlock,ImpliedFactorReturnsState-method}
\alias{setImpliedFactorReturnsDataObject,VirtualAnalysisBlock,VirtualImpliedFactorReturnsData-method}
\title{Set implied_factor_returns object in object slot}
\usage{

  \S4method{setImpliedFactorReturnsDataObject}{PortfolioFactorStateProbCondLossAnalysisBlock,ImpliedFactorReturnsState}(object,
  implied_factor_returns)
}
\arguments{
\item{object}{object of class derived from"VirtualAnalysisBlock"}

\item{implied_factor_returns}{object of class derived from "VirtualImpliedFactorReturnsData"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"
}
\description{
Public method to set implied_factor_returns slot with
class derived from  "VirtualImpliedFactorReturnsData"
class object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = PortfolioFactorStateProbCondLossAnalysisBlock,implied_factor_returns = ImpliedFactorReturnsState}: Set portfolio object in object slot
}}

