% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_returns.r
\docType{methods}
\name{setInstrumentBetasDataObject,PortfolioCoreFactorReturnsAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,PortfolioCoreFactorReturnsAnalysisBlock,InstrumentBetasData-method}
\title{Set instrument_betas object in object slot}
\usage{

  \S4method{setInstrumentBetasDataObject}{PortfolioCoreFactorReturnsAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)
}
\arguments{
\item{object}{object of class "PortfolioCoreFactorReturnsAnalysisBlock"}

\item{instrument_betas}{object of class "InstrumentBetasData"}
}
\value{
\code{object} object of class "PortfolioCoreFactorReturnsAnalysisBlock"
}
\description{
Public method to set instrument_betas slot with "InstrumentBetasData"
class object
}

