% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setInstrumentBetasDataObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,InstrumentBetasData-method}
\title{Set instrument_betas object in object slot}
\usage{

  \S4method{setInstrumentBetasDataObject}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)
}
\arguments{
\item{object}{object of class "PortfolioInstrumentCoreFactorsMCTRAnalysisBlock"}

\item{instrument_betas}{object of class "InstrumentBetasData"}
}
\value{
\code{object} object of class "PortfolioInstrumentCoreFactorsMCTRAnalysisBlock"
}
\description{
Public method to set instrument_betas slot with "InstrumentBetasData"
class object
}

