% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk.r
\docType{methods}
\name{setInstrumentBetasDataObject,PortfolioInstrumentMCTRAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,PortfolioInstrumentMCTRAnalysisBlock,InstrumentBetasData-method}
\title{Set instrument_betas object in object slot}
\usage{

  \S4method{setInstrumentBetasDataObject}{PortfolioInstrumentMCTRAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)
}
\arguments{
\item{object}{object of class "PortfolioInstrumentMCTRAnalysisBlock"}

\item{instrument_betas}{object of class "InstrumentBetasData"}
}
\value{
\code{object} object of class "PortfolioInstrumentMCTRAnalysisBlock"
}
\description{
Public method to set instrument_betas slot with "InstrumentBetasData"
class object
}

