% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/portfolio_core_factor_exposures.r, R/portfolio_core_factor_hedge_exposures.r, R/portfolio_core_factor_hedge_returns.r, R/portfolio_core_factor_returns.r, R/portfolio_factor_exposures.r, R/portfolio_factor_returns.r, R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r, R/portfolio_instrument_marginal_contribution_to_risk.r, R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setInstrumentBetasDataObject}
\alias{setInstrumentBetasDataObject}
\alias{setInstrumentBetasDataObject,PortfolioCoreFactorExposuresAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,PortfolioCoreFactorHedgeExposuresAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,PortfolioCoreFactorHedgeReturnsAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,PortfolioCoreFactorReturnsAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,PortfolioFactorExposuresAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,PortfolioFactorReturnsAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,PortfolioInstrumentMCTRAnalysisBlock,InstrumentBetasData-method}
\alias{setInstrumentBetasDataObject,VirtualAnalysisBlock,InstrumentBetasData-method}
\title{Set instrument_betas object in object slot}
\usage{

  \S4method{setInstrumentBetasDataObject}{PortfolioCoreFactorExposuresAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)


  \S4method{setInstrumentBetasDataObject}{PortfolioCoreFactorHedgeExposuresAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)


  \S4method{setInstrumentBetasDataObject}{PortfolioCoreFactorHedgeReturnsAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)


  \S4method{setInstrumentBetasDataObject}{PortfolioCoreFactorReturnsAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)


  \S4method{setInstrumentBetasDataObject}{PortfolioFactorExposuresAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)


  \S4method{setInstrumentBetasDataObject}{PortfolioFactorReturnsAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)


  \S4method{setInstrumentBetasDataObject}{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)


  \S4method{setInstrumentBetasDataObject}{PortfolioInstrumentMCTRAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)


  \S4method{setInstrumentBetasDataObject}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,InstrumentBetasData}(object,
  instrument_betas)
}
\arguments{
\item{object}{object of class derived from"VirtualAnalysisBlock"}

\item{instrument_betas}{object of class "InstrumentBetasData"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"
}
\description{
Public method to set instrument_betas slot with "InstrumentBetasData"
class object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = PortfolioCoreFactorExposuresAnalysisBlock,instrument_betas = InstrumentBetasData}: Set instrument_betas object in object slot

\item \code{object = PortfolioCoreFactorHedgeExposuresAnalysisBlock,instrument_betas = InstrumentBetasData}: Set portfolio object in object slot

\item \code{object = PortfolioCoreFactorHedgeReturnsAnalysisBlock,instrument_betas = InstrumentBetasData}: Set portfolio object in object slot

\item \code{object = PortfolioCoreFactorReturnsAnalysisBlock,instrument_betas = InstrumentBetasData}: Set portfolio object in object slot

\item \code{object = PortfolioFactorExposuresAnalysisBlock,instrument_betas = InstrumentBetasData}: Set portfolio object in object slot

\item \code{object = PortfolioFactorReturnsAnalysisBlock,instrument_betas = InstrumentBetasData}: Set portfolio object in object slot

\item \code{object = PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,instrument_betas = InstrumentBetasData}: Set portfolio object in object slot

\item \code{object = PortfolioInstrumentMCTRAnalysisBlock,instrument_betas = InstrumentBetasData}: Set instrument betas object in object slot

\item \code{object = PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,instrument_betas = InstrumentBetasData}: Set instrument betas in object slot
}}

