% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buys_and_sells.r
\docType{methods}
\name{setMarketDataObject,BuysAndSellsAnalysisBlock,MarketData-method}
\alias{setMarketDataObject,BuysAndSellsAnalysisBlock,MarketData-method}
\title{Set market_data object in object slot}
\usage{
\S4method{setMarketDataObject}{BuysAndSellsAnalysisBlock,MarketData}(object,
  market_data)
}
\arguments{
\item{object}{object of class "BuysAndSellsAnalysisBlock"}

\item{market_data}{object of class "MarketData"}
}
\value{
\code{object} object of class "BuysAndSellsAnalysisBlock"
}
\description{
Public method to set market_data slot with "MarketData"
class object
}

