% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_market_return.r
\docType{methods}
\name{setMarketDataObject,RelativeMarketReturnAnalysisBlock,MarketDataSX5E-method}
\alias{setMarketDataObject,RelativeMarketReturnAnalysisBlock,MarketDataSX5E-method}
\title{Set market_data object in object slot}
\usage{

  \S4method{setMarketDataObject}{RelativeMarketReturnAnalysisBlock,MarketDataSX5E}(object,
  market_data)
}
\arguments{
\item{object}{object of class "RelativeMarketReturnAnalysisBlock"}

\item{market_data}{object of class "MarketDataSX5E"}
}
\value{
\code{object} object of class "RelativeMarketReturnAnalysisBlock"
}
\description{
Public method to set market_data slot with "MarketDataSX5E"
class object
}

