% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_hedge_exposures.r
\docType{methods}
\name{setPortfolioDataObject,IndexPortfolioCoreFactorHedgeExposuresAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioCoreFactorHedgeExposuresAnalysisBlock,VirtualIndexPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{IndexPortfolioCoreFactorHedgeExposuresAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class "IndexPortfolioCoreFactorHedgeExposuresAnalysisBlock"}

\item{portfolio}{object of class "VirtualIndexPortfolio"}
}
\value{
\code{object} object of class "IndexPortfolioCoreFactorHedgeExposuresAnalysisBlock"
}
\description{
Public method to set portfolio slot with "VirtualIndexPortfolio"
class object
}

