% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_returns.r
\docType{methods}
\name{setPortfolioDataObject,IndexPortfolioCoreFactorReturnsAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioCoreFactorReturnsAnalysisBlock,VirtualIndexPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{IndexPortfolioCoreFactorReturnsAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class "IndexPortfolioCoreFactorReturnsAnalysisBlock"}

\item{portfolio}{object of class "VirtualIndexPortfolio"}
}
\value{
\code{object} object of class "IndexPortfolioCoreFactorReturnsAnalysisBlock"
}
\description{
Public method to set portfolio slot with "VirtualIndexPortfolio"
class object
}

