% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_exposures.r
\docType{methods}
\name{setPortfolioDataObject,IndexPortfolioFactorExposuresAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioFactorExposuresAnalysisBlock,VirtualIndexPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{IndexPortfolioFactorExposuresAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class "IndexPortfolioFactorExposuresAnalysisBlock"}

\item{portfolio}{object of class "VirtualIndexPortfolio"}
}
\value{
\code{object} object of class "IndexPortfolioFactorExposuresAnalysisBlock"
}
\description{
Public method to set portfolio slot with "VirtualIndexPortfolio"
class object
}

