% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setPortfolioDataObject,IndexPortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,VirtualIndexPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{IndexPortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class "IndexPortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock"}

\item{portfolio}{object of class "VirtualIndexPortfolio"}
}
\value{
\code{object} object of class "IndexPortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock""
}
\description{
Public method to set portfolio slot with "VirtualIndexPortfolio"
class object
}

