% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setPortfolioDataObject,IndexPortfolioInstrumentCoreFactorsMCTRAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioInstrumentCoreFactorsMCTRAnalysisBlock,VirtualIndexPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{IndexPortfolioInstrumentCoreFactorsMCTRAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class "IndexPortfolioInstrumentCoreFactorsMCTRAnalysisBlock"}

\item{portfolio}{object of class "VirtualIndexPortfolio"}
}
\value{
\code{object} object of class "IndexPortfolioInstrumentCoreFactorsMCTRAnalysisBlock""
}
\description{
Public method to set portfolio slot with "VirtualIndexPortfolio"
class object
}

