% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{methods}
\name{setPortfolioDataObject,IndexPortfolioVarianceDecompositionAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioVarianceDecompositionAnalysisBlock,VirtualIndexPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{IndexPortfolioVarianceDecompositionAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class "IndexPortfolioVarianceDecompositionAnalysisBlock"}

\item{portfolio}{object of class "VirtualIndexPortfolio"}
}
\value{
\code{object} object of class "IndexPortfolioVarianceDecompositionAnalysisBlock""
}
\description{
Public method to set portfolio slot with "VirtualIndexPortfolio"
class object
}

