% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_exposures.r
\docType{methods}
\name{setPortfolioDataObject,PortfolioCoreFactorExposuresAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioCoreFactorExposuresAnalysisBlock,StrategyPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{PortfolioCoreFactorExposuresAnalysisBlock,StrategyPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class "PortfolioCoreFactorExposuresAnalysisBlock"}

\item{portfolio}{object of class "StrategyPortfolio"}
}
\value{
\code{object} object of class "PortfolioCoreFactorExposuresAnalysisBlock"
}
\description{
Public method to set portfolio slot with "StrategyPortfolio"
class object
}

