% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_hedge_returns.r
\docType{methods}
\name{setPortfolioDataObject,PortfolioCoreFactorHedgeReturnsAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioCoreFactorHedgeReturnsAnalysisBlock,StrategyPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{PortfolioCoreFactorHedgeReturnsAnalysisBlock,StrategyPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class "PortfolioCoreFactorHedgeReturnsAnalysisBlock"}

\item{portfolio}{object of class "StrategyPortfolio"}
}
\value{
\code{object} object of class "PortfolioCoreFactorHedgeReturnsAnalysisBlock"
}
\description{
Public method to set portfolio slot with "StrategyPortfolio"
class object
}

