% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_exposures.r
\docType{methods}
\name{setPortfolioDataObject,PortfolioFactorExposuresAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioFactorExposuresAnalysisBlock,StrategyPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{PortfolioFactorExposuresAnalysisBlock,StrategyPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class "PortfolioFactorExposuresAnalysisBlock"}

\item{portfolio}{object of class "StrategyPortfolio"}
}
\value{
\code{object} object of class "PortfolioFactorExposuresAnalysisBlock"
}
\description{
Public method to set portfolio slot with "StrategyPortfolio"
class object
}

