% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setPortfolioDataObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,StrategyPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,StrategyPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class "PortfolioInstrumentCoreFactorsMCTRAnalysisBlock"}

\item{portfolio}{object of class "StrategyPortfolio"}
}
\value{
\code{object} object of class "PortfolioInstrumentCoreFactorsMCTRAnalysisBlock"
}
\description{
Public method to set portfolio slot with "StrategyPortfolio"
class object
}

