% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/portfolio_variance_decomposition.r, R/portfolio_core_factor_exposures.r, R/portfolio_core_factor_hedge_exposures.r, R/portfolio_core_factor_hedge_returns.r, R/portfolio_core_factor_returns.r, R/portfolio_factor_exposures.r, R/portfolio_factor_returns.r, R/portfolio_factor_state_prob_cond_loss.r, R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r, R/portfolio_instrument_marginal_contribution_to_risk.r, R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setPortfolioDataObject}
\alias{setPortfolioDataObject}
\alias{setPortfolioDataObject,IndexPortfolioCoreFactorExposuresAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioCoreFactorHedgeExposuresAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioCoreFactorHedgeReturnsAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioCoreFactorReturnsAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioFactorExposuresAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioFactorReturnsAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioInstrumentCoreFactorsMCTRAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioInstrumentMCTRAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioVarianceDecompositionAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,PortfolioCoreFactorExposuresAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioCoreFactorHedgeExposuresAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioCoreFactorHedgeReturnsAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioCoreFactorReturnsAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioFactorExposuresAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioFactorReturnsAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioFactorStateProbCondLossAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,PortfolioInstrumentMCTRAnalysisBlock,StrategyPortfolio-method}
\alias{setPortfolioDataObject,VirtualAnalysisBlock,VirtualPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{IndexPortfolioVarianceDecompositionAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{PortfolioCoreFactorExposuresAnalysisBlock,StrategyPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{IndexPortfolioCoreFactorExposuresAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{PortfolioCoreFactorHedgeExposuresAnalysisBlock,StrategyPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{IndexPortfolioCoreFactorHedgeExposuresAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{PortfolioCoreFactorHedgeReturnsAnalysisBlock,StrategyPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{IndexPortfolioCoreFactorHedgeReturnsAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{PortfolioCoreFactorReturnsAnalysisBlock,StrategyPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{IndexPortfolioCoreFactorReturnsAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{PortfolioFactorExposuresAnalysisBlock,StrategyPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{IndexPortfolioFactorExposuresAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{PortfolioFactorReturnsAnalysisBlock,StrategyPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{IndexPortfolioFactorReturnsAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{PortfolioFactorStateProbCondLossAnalysisBlock,StrategyPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,StrategyPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{IndexPortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{PortfolioInstrumentMCTRAnalysisBlock,StrategyPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{IndexPortfolioInstrumentMCTRAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,StrategyPortfolio}(object,
  portfolio)


  \S4method{setPortfolioDataObject}{IndexPortfolioInstrumentCoreFactorsMCTRAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class derived from"VirtualAnalysisBlock"}

\item{portfolio}{object of class "StrategyPortfolio" or "VirtualIndexPortfolio"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"
}
\description{
Public method to set portfolio slot with "StrategyPortfolio"
class object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = IndexPortfolioVarianceDecompositionAnalysisBlock,portfolio = VirtualIndexPortfolio}: Set portfolio object in object slot

\item \code{object = PortfolioCoreFactorExposuresAnalysisBlock,portfolio = StrategyPortfolio}: Set portfolio object in object slot

\item \code{object = IndexPortfolioCoreFactorExposuresAnalysisBlock,portfolio = VirtualIndexPortfolio}: Set portfolio object in object slot

\item \code{object = PortfolioCoreFactorHedgeExposuresAnalysisBlock,portfolio = StrategyPortfolio}: Set portfolio object in object slot

\item \code{object = IndexPortfolioCoreFactorHedgeExposuresAnalysisBlock,portfolio = VirtualIndexPortfolio}: Set portfolio object in object slot

\item \code{object = PortfolioCoreFactorHedgeReturnsAnalysisBlock,portfolio = StrategyPortfolio}: Set portfolio object in object slot

\item \code{object = IndexPortfolioCoreFactorHedgeReturnsAnalysisBlock,portfolio = VirtualIndexPortfolio}: Set portfolio object in object slot

\item \code{object = PortfolioCoreFactorReturnsAnalysisBlock,portfolio = StrategyPortfolio}: Set portfolio object in object slot

\item \code{object = IndexPortfolioCoreFactorReturnsAnalysisBlock,portfolio = VirtualIndexPortfolio}: Set portfolio object in object slot

\item \code{object = PortfolioFactorExposuresAnalysisBlock,portfolio = StrategyPortfolio}: Set portfolio object in object slot

\item \code{object = IndexPortfolioFactorExposuresAnalysisBlock,portfolio = VirtualIndexPortfolio}: Set portfolio object in object slot

\item \code{object = PortfolioFactorReturnsAnalysisBlock,portfolio = StrategyPortfolio}: Set portfolio object in object slot

\item \code{object = IndexPortfolioFactorReturnsAnalysisBlock,portfolio = VirtualIndexPortfolio}: Set portfolio object in object slot

\item \code{object = PortfolioFactorStateProbCondLossAnalysisBlock,portfolio = StrategyPortfolio}: Set portfolio object in object slot

\item \code{object = PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,portfolio = StrategyPortfolio}: Set portfolio object in object slot

\item \code{object = IndexPortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,portfolio = VirtualIndexPortfolio}: Set portfolio object in object slot

\item \code{object = PortfolioInstrumentMCTRAnalysisBlock,portfolio = StrategyPortfolio}: Set portfolio object in object slot

\item \code{object = IndexPortfolioInstrumentMCTRAnalysisBlock,portfolio = VirtualIndexPortfolio}: Set portfolio object in object slot

\item \code{object = PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,portfolio = StrategyPortfolio}: Set portfolio object in object slot

\item \code{object = IndexPortfolioInstrumentCoreFactorsMCTRAnalysisBlock,portfolio = VirtualIndexPortfolio}: Set portfolio object in object slot
}}

