% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades.r
\docType{methods}
\name{setPositionDataObject,AverageDownTradesAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,AverageDownTradesAnalysisBlock,OffsidePositionData-method}
\title{Set position_data object in object slot}
\usage{

  \S4method{setPositionDataObject}{AverageDownTradesAnalysisBlock,OffsidePositionData}(object,
  position_data)
}
\arguments{
\item{object}{object of class "AverageDownTradesAnalysisBlock"}

\item{position_data}{object of class "OffsidePositionData"}
}
\value{
\code{object} object of class "AverageDownTradesAnalysisBlock"
}
\description{
Public method to set position_data slot with "OffsidePositionData"
class object
}

