% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_relative_bps_per_month.r
\docType{methods}
\name{setPositionDataObject,OffsidePositionsBpsPerMonthAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,OffsidePositionsBpsPerMonthAnalysisBlock,OffsidePositionData-method}
\title{Set position_data object in object slot}
\usage{

  \S4method{setPositionDataObject}{OffsidePositionsBpsPerMonthAnalysisBlock,OffsidePositionData}(object,
  position_data)
}
\arguments{
\item{object}{object of class "OffsidePositionsBpsPerMonthAnalysisBlock"}

\item{position_data}{object of class "OffsidePositionData"}
}
\value{
\code{object} object of class "OffsidePositionsBpsPerMonthAnalysisBlock"
}
\description{
Public method to set position_data slot with "OffsidePositionData"
class object
}

