% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_cumulative_pnl.r
\docType{methods}
\name{setPositionDataObject,OffsidePositionsCumulativePnLAnalysisBlock,OffsidePositionGainData-method}
\alias{setPositionDataObject,OffsidePositionsCumulativePnLAnalysisBlock,OffsidePositionGainData-method}
\title{Set position_data object in object slot}
\usage{

  \S4method{setPositionDataObject}{OffsidePositionsCumulativePnLAnalysisBlock,OffsidePositionGainData}(object,
  position_data)
}
\arguments{
\item{object}{object of class "OffsidePositionsCumulativePnLAnalysisBlock"}

\item{position_data}{object of class "OffsidePositionGainData"}
}
\value{
\code{object} object of class "OffsidePositionsCumulativePnLAnalysisBlock"
}
\description{
Public method to set position_data slot with "OffsidePositionGainData"
class object
}

