% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_gain_vs_days.r
\docType{methods}
\name{setPositionDataObject,OffsidePositionsGainVsDaysAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,OffsidePositionsGainVsDaysAnalysisBlock,OffsidePositionData-method}
\title{Set position_data object in object slot}
\usage{

  \S4method{setPositionDataObject}{OffsidePositionsGainVsDaysAnalysisBlock,OffsidePositionData}(object,
  position_data)
}
\arguments{
\item{object}{object of class "OffsidePositionsGainVsDaysAnalysisBlock"}

\item{position_data}{object of class "OffsidePositionData"}
}
\value{
\code{object} object of class "OffsidePositionsGainVsDaysAnalysisBlock"
}
\description{
Public method to set position_data slot with "OffsidePositionData"
class object
}

