% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offside_positions_pnl_vs_days.r
\docType{methods}
\name{setPositionDataObject,OffsidePositionsPnLVsDaysAnalysisBlock,OffsidePositionGainData-method}
\alias{setPositionDataObject,OffsidePositionsPnLVsDaysAnalysisBlock,OffsidePositionGainData-method}
\title{Set position_data object in object slot}
\usage{

  \S4method{setPositionDataObject}{OffsidePositionsPnLVsDaysAnalysisBlock,OffsidePositionGainData}(object,
  position_data)
}
\arguments{
\item{object}{object of class "OffsidePositionsPnLVsDaysAnalysisBlock"}

\item{position_data}{object of class "OffsidePositionGainData"}
}
\value{
\code{object} object of class "OffsidePositionsPnLVsDaysAnalysisBlock"
}
\description{
Public method to set position_data slot with "OffsidePositionGainData"
class object
}

