% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_day_zero_pnl.r
\docType{methods}
\name{setPositionDataObject,PositionsHoldingDayZeroPnLAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,PositionsHoldingDayZeroPnLAnalysisBlock,OffsidePositionData-method}
\title{Set position_data object in object slot}
\usage{

  \S4method{setPositionDataObject}{PositionsHoldingDayZeroPnLAnalysisBlock,OffsidePositionData}(object,
  position_data)
}
\arguments{
\item{object}{object of class "PositionsHoldingDayZeroPnLAnalysisBlock"}

\item{position_data}{object of class "OffsidePositionData"}
}
\value{
\code{object} object of class "PositionsHoldingDayZeroPnLAnalysisBlock"
}
\description{
Public method to set position_data slot with "OffsidePositionData"
class object
}

