% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades_performance_on_results_day.r
\docType{methods}
\name{setPositionDataObject,TradesPerformanceOnResultsDayAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,TradesPerformanceOnResultsDayAnalysisBlock,OffsidePositionData-method}
\title{Set position_data object in object slot}
\usage{

  \S4method{setPositionDataObject}{TradesPerformanceOnResultsDayAnalysisBlock,OffsidePositionData}(object,
  position_data)
}
\arguments{
\item{object}{object of class "TradesPerformanceOnResultsDayAnalysisBlock"}

\item{position_data}{object of class "OffsidePositionData"}
}
\value{
\code{object} object of class "TradesPerformanceOnResultsDayAnalysisBlock"
}
\description{
Public method to set position_data slot with "OffsidePositionData"
class object
}

