% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/average_down_trades.r, R/offside_positions_gain_vs_days.r, R/offside_positions_cumulative_pnl.r, R/offside_positions_pnl_vs_days.r, R/offside_positions_relative_bps_per_month.r, R/positions_holding_capital_distribution.r, R/positions_holding_day_zero_pnl.r, R/positions_holding_period.r, R/trades_performance_on_results_day.r
\docType{methods}
\name{setPositionDataObject}
\alias{setPositionDataObject}
\alias{setPositionDataObject,AverageDownTradesAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,OffsidePositionsBpsPerMonthAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,OffsidePositionsCumulativePnLAnalysisBlock,OffsidePositionGainData-method}
\alias{setPositionDataObject,OffsidePositionsGainVsDaysAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,OffsidePositionsPnLVsDaysAnalysisBlock,OffsidePositionGainData-method}
\alias{setPositionDataObject,PositionsHoldingCapitalDistributionAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,PositionsHoldingDayZeroPnLAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,PositionsHoldingPeriodAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,TradesPerformanceOnResultsDayAnalysisBlock,OffsidePositionData-method}
\alias{setPositionDataObject,VirtualAnalysisBlock,VirtualPositionData-method}
\title{Set position_data object in object slot}
\usage{

  \S4method{setPositionDataObject}{AverageDownTradesAnalysisBlock,OffsidePositionData}(object,
  position_data)


  \S4method{setPositionDataObject}{OffsidePositionsGainVsDaysAnalysisBlock,OffsidePositionData}(object,
  position_data)


  \S4method{setPositionDataObject}{OffsidePositionsCumulativePnLAnalysisBlock,OffsidePositionGainData}(object,
  position_data)


  \S4method{setPositionDataObject}{OffsidePositionsPnLVsDaysAnalysisBlock,OffsidePositionGainData}(object,
  position_data)


  \S4method{setPositionDataObject}{OffsidePositionsBpsPerMonthAnalysisBlock,OffsidePositionData}(object,
  position_data)


  \S4method{setPositionDataObject}{PositionsHoldingCapitalDistributionAnalysisBlock,OffsidePositionData}(object,
  position_data)


  \S4method{setPositionDataObject}{PositionsHoldingDayZeroPnLAnalysisBlock,OffsidePositionData}(object,
  position_data)


  \S4method{setPositionDataObject}{PositionsHoldingPeriodAnalysisBlock,OffsidePositionData}(object,
  position_data)


  \S4method{setPositionDataObject}{TradesPerformanceOnResultsDayAnalysisBlock,OffsidePositionData}(object,
  position_data)
}
\arguments{
\item{object}{object of class derived from"VirtualAnalysisBlock"}

\item{position_data}{object of class "VirtualPositionData"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"
}
\description{
Public method to set position_data slot with
object of class derived from "VirtualPositionData"
}
\section{Methods (by class)}{
\itemize{
\item \code{object = AverageDownTradesAnalysisBlock,position_data = OffsidePositionData}: Set position_data object in object slot

\item \code{object = OffsidePositionsGainVsDaysAnalysisBlock,position_data = OffsidePositionData}: Set position_data object in object slot

\item \code{object = OffsidePositionsCumulativePnLAnalysisBlock,position_data = OffsidePositionGainData}: Set position_data object in object slot

\item \code{object = OffsidePositionsPnLVsDaysAnalysisBlock,position_data = OffsidePositionGainData}: Set position_data object in object slot

\item \code{object = OffsidePositionsBpsPerMonthAnalysisBlock,position_data = OffsidePositionData}: Set position_data object in object slot

\item \code{object = PositionsHoldingCapitalDistributionAnalysisBlock,position_data = OffsidePositionData}: Set position_data object in object slot

\item \code{object = PositionsHoldingDayZeroPnLAnalysisBlock,position_data = OffsidePositionData}: Set position_data object in object slot

\item \code{object = PositionsHoldingPeriodAnalysisBlock,position_data = OffsidePositionData}: Set position_data object in object slot

\item \code{object = TradesPerformanceOnResultsDayAnalysisBlock,position_data = OffsidePositionData}: Set position_data object in object slot
}}

