% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades_performance_on_results_day.r
\docType{methods}
\name{setPriceDataObject,TradesPerformanceOnResultsDayAnalysisBlock,PriceData-method}
\alias{setPriceDataObject,TradesPerformanceOnResultsDayAnalysisBlock,PriceData-method}
\title{Set price_data object in object slot}
\usage{

  \S4method{setPriceDataObject}{TradesPerformanceOnResultsDayAnalysisBlock,PriceData}(object,
  price_data)
}
\arguments{
\item{object}{object of class "TradesPerformanceOnResultsDayAnalysisBlock"}

\item{price_data}{object of class "PriceData"}
}
\value{
\code{object} object of class "TradesPerformanceOnResultsDayAnalysisBlock"
}
\description{
Public method to set price_data slot with "PriceData"
class object
}

