% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_state.r
\docType{methods}
\name{setRiskModelObject,ImpliedFactorReturnsStateAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,ImpliedFactorReturnsStateAnalysisBlock,VirtualRiskModel-method}
\title{Set risk_model object in object slot}
\usage{

  \S4method{setRiskModelObject}{ImpliedFactorReturnsStateAnalysisBlock,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "ImpliedFactorReturnsStateAnalysisBlock"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{object} object of class "ImpliedFactorReturnsStateAnalysisBlock"
}
\description{
Public method to set trade_data slot with "VirtualRiskModel"
class object
}

