% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_timeseries.r
\docType{methods}
\name{setRiskModelObject,MarketStyleAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,MarketStyleAnalysisBlock,VirtualRiskModel-method}
\title{Set risk_model object in object slot}
\usage{

  \S4method{setRiskModelObject}{MarketStyleAnalysisBlock,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "MarketStyleAnalysisBlock"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{object} object of class "MarketStyleAnalysisBlock"
}
\description{
Public method to set trade_data slot with "VirtualRiskModel"
class object
}

