% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_factor_stat.r
\docType{methods}
\name{setRiskModelObject,MarketStyleFactorStatisticAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,MarketStyleFactorStatisticAnalysisBlock,VirtualRiskModel-method}
\title{Set risk_model object in object slot}
\usage{

  \S4method{setRiskModelObject}{MarketStyleFactorStatisticAnalysisBlock,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "MarketStyleFactorStatisticAnalysisBlock"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{object} object of class "MarketStyleFactorStatisticAnalysisBlock"
}
\description{
Public method to set trade_data slot with "VirtualRiskModel"
class object
}

