% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_hedge_exposures.r
\docType{methods}
\name{setRiskModelObject,PortfolioCoreFactorHedgeExposuresAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioCoreFactorHedgeExposuresAnalysisBlock,VirtualRiskModel-method}
\title{Set risk_model object in object slot}
\usage{

  \S4method{setRiskModelObject}{PortfolioCoreFactorHedgeExposuresAnalysisBlock,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "PortfolioCoreFactorHedgeExposuresAnalysisBlock"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{object} object of class "PortfolioCoreFactorHedgeExposuresAnalysisBlock"
}
\description{
Public method to set trade_data slot with "VirtualRiskModel"
class object
}

