% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_core_factor_returns.r
\docType{methods}
\name{setRiskModelObject,PortfolioCoreFactorReturnsAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioCoreFactorReturnsAnalysisBlock,VirtualRiskModel-method}
\title{Set risk_model object in object slot}
\usage{

  \S4method{setRiskModelObject}{PortfolioCoreFactorReturnsAnalysisBlock,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "PortfolioCoreFactorReturnsAnalysisBlock"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{object} object of class "PortfolioCoreFactorReturnsAnalysisBlock"
}
\description{
Public method to set trade_data slot with "VirtualRiskModel"
class object
}

