% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_factor_exposures.r
\docType{methods}
\name{setRiskModelObject,PortfolioFactorExposuresAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioFactorExposuresAnalysisBlock,VirtualRiskModel-method}
\title{Set risk_model object in object slot}
\usage{

  \S4method{setRiskModelObject}{PortfolioFactorExposuresAnalysisBlock,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "PortfolioFactorExposuresAnalysisBlock"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{object} object of class "PortfolioFactorExposuresAnalysisBlock"
}
\description{
Public method to set trade_data slot with "VirtualRiskModel"
class object
}

