% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setRiskModelObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,VirtualRiskModel-method}
\title{Set risk_model object in object slot}
\usage{

  \S4method{setRiskModelObject}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "PortfolioInstrumentCoreFactorsMCTRAnalysisBlock"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{object} object of class "PortfolioInstrumentCoreFactorsMCTRAnalysisBlock"
}
\description{
Public method to set trade_data slot with "VirtualRiskModel"
class object
}

