% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_variance_decomposition.r
\docType{methods}
\name{setRiskModelObject,PortfolioVarianceDecompositionAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioVarianceDecompositionAnalysisBlock,VirtualRiskModel-method}
\title{Set risk_model object in object slot}
\usage{

  \S4method{setRiskModelObject}{PortfolioVarianceDecompositionAnalysisBlock,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "PortfolioVarianceDecompositionAnalysisBlock"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{object} object of class "PortfolioVarianceDecompositionAnalysisBlock"
}
\description{
Public method to set trade_data slot with "VirtualRiskModel"
class object
}

