% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/implied_factor_returns.r, R/implied_factor_returns_state.r, R/portfolio_variance_decomposition.r, R/market_style_factor_stat.r, R/market_style_timeseries.r, R/portfolio_core_factor_exposures.r, R/portfolio_core_factor_hedge_exposures.r, R/portfolio_core_factor_hedge_returns.r, R/portfolio_core_factor_returns.r, R/portfolio_factor_exposures.r, R/portfolio_factor_returns.r, R/portfolio_factor_state_prob_cond_loss.r, R/portfolio_hedge_instrument_marginal_contribution_to_risk_core_factors.r, R/portfolio_instrument_marginal_contribution_to_risk.r, R/portfolio_instrument_marginal_contribution_to_risk_core_factors.r
\docType{methods}
\name{setRiskModelObject}
\alias{setRiskModelObject}
\alias{setRiskModelObject,ImpliedFactorReturnsAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,ImpliedFactorReturnsStateAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,MarketStyleAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,MarketStyleFactorStatisticAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioCoreFactorExposuresAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioCoreFactorHedgeExposuresAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioCoreFactorHedgeReturnsAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioCoreFactorReturnsAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioFactorExposuresAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioFactorReturnsAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioFactorStateProbCondLossAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioInstrumentMCTRAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,PortfolioVarianceDecompositionAnalysisBlock,VirtualRiskModel-method}
\alias{setRiskModelObject,VirtualAnalysisBlock,VirtualRiskModel-method}
\title{Set risk_model object in object slot}
\usage{

  \S4method{setRiskModelObject}{ImpliedFactorReturnsAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{ImpliedFactorReturnsStateAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{PortfolioVarianceDecompositionAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{MarketStyleFactorStatisticAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{MarketStyleAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{PortfolioCoreFactorExposuresAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{PortfolioCoreFactorHedgeExposuresAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{PortfolioCoreFactorHedgeReturnsAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{PortfolioCoreFactorReturnsAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{PortfolioFactorExposuresAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{PortfolioFactorReturnsAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{PortfolioFactorStateProbCondLossAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{PortfolioInstrumentMCTRAnalysisBlock,VirtualRiskModel}(object,
  risk_model)


  \S4method{setRiskModelObject}{PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class derived from "VirtualAnalysisBlock"}

\item{risk_model}{object of class derived from "VirtualRiskModel"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"
}
\description{
Public method to set risk_model slot with "VirtualRiskModel"
class object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = ImpliedFactorReturnsAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = ImpliedFactorReturnsStateAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = PortfolioVarianceDecompositionAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = MarketStyleFactorStatisticAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = MarketStyleAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = PortfolioCoreFactorExposuresAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = PortfolioCoreFactorHedgeExposuresAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = PortfolioCoreFactorHedgeReturnsAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = PortfolioCoreFactorReturnsAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = PortfolioFactorExposuresAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = PortfolioFactorReturnsAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = PortfolioFactorStateProbCondLossAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = PortfolioHedgeInstrumentCoreFactorsMCTRAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = PortfolioInstrumentMCTRAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot

\item \code{object = PortfolioInstrumentCoreFactorsMCTRAnalysisBlock,risk_model = VirtualRiskModel}: Set risk_model object in object slot
}}

