% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_aum_and_turnover.r
\docType{methods}
\name{setStrategyDataObject,StrategyBreakdownAUMAndTurnoverAnalysisBlock,StrategyBreakDownData-method}
\alias{setStrategyDataObject,StrategyBreakdownAUMAndTurnoverAnalysisBlock,StrategyBreakDownData-method}
\title{Set strategy_data object in object slot}
\usage{

  \S4method{setStrategyDataObject}{StrategyBreakdownAUMAndTurnoverAnalysisBlock,StrategyBreakDownData}(object,
  strategy_data)
}
\arguments{
\item{object}{object of class "StrategyBreakdownAUMAndTurnoverAnalysisBlock"}

\item{strategy_data}{object of class "StrategyBreakDownData"}
}
\value{
\code{object} object of class "StrategyBreakdownAUMAndTurnoverAnalysisBlock"
}
\description{
Public method to set strategy_data slot with "StrategyBreakDownData"
class object
}

