% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_total_pnl_and_pnl_delta.r
\docType{methods}
\name{setStrategyDataObject,StrategyBreakdownTotalAndDeltaPnLAnalysisBlock,StrategyBreakDownData-method}
\alias{setStrategyDataObject,StrategyBreakdownTotalAndDeltaPnLAnalysisBlock,StrategyBreakDownData-method}
\title{Set strategy_data object in object slot}
\usage{

  \S4method{setStrategyDataObject}{StrategyBreakdownTotalAndDeltaPnLAnalysisBlock,StrategyBreakDownData}(object,
  strategy_data)
}
\arguments{
\item{object}{object of class "StrategyBreakdownTotalAndDeltaPnLAnalysisBlock"}

\item{strategy_data}{object of class "StrategyBreakDownData"}
}
\value{
\code{object} object of class "StrategyBreakdownTotalAndDeltaPnLAnalysisBlock"
}
\description{
Public method to set strategy_data slot with "StrategyBreakDownData"
class object
}

