% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/strategy_breakdown_aum_and_turnover.r, R/strategy_breakdown_total_pnl_and_pnl_delta.r
\docType{methods}
\name{setStrategyDataObject}
\alias{setStrategyDataObject}
\alias{setStrategyDataObject,StrategyBreakdownAUMAndTurnoverAnalysisBlock,StrategyBreakDownData-method}
\alias{setStrategyDataObject,StrategyBreakdownTotalAndDeltaPnLAnalysisBlock,StrategyBreakDownData-method}
\alias{setStrategyDataObject,VirtualAnalysisBlock,VirtualStrategyData-method}
\title{Set strategy_data object in object slot}
\usage{

  \S4method{setStrategyDataObject}{StrategyBreakdownAUMAndTurnoverAnalysisBlock,StrategyBreakDownData}(object,
  strategy_data)


  \S4method{setStrategyDataObject}{StrategyBreakdownTotalAndDeltaPnLAnalysisBlock,StrategyBreakDownData}(object,
  strategy_data)
}
\arguments{
\item{object}{object of class derived from"VirtualAnalysisBlock"}

\item{strategy_data}{object of class derived from "VirtualStrategyData"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"
}
\description{
Public method to set strategy_data slot with "VirtualStrategyData"
class or derived object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = StrategyBreakdownAUMAndTurnoverAnalysisBlock,strategy_data = StrategyBreakDownData}: Set strategy_data object in object slot

\item \code{object = StrategyBreakdownTotalAndDeltaPnLAnalysisBlock,strategy_data = StrategyBreakDownData}: Set strategy_data object in object slot
}}

