% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades_focus.r
\docType{methods}
\name{setTradeDataObject,AverageDownTradesFocusAnalysisBlock,AverageDownTradesData-method}
\alias{setTradeDataObject,AverageDownTradesFocusAnalysisBlock,AverageDownTradesData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{AverageDownTradesFocusAnalysisBlock,AverageDownTradesData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "AverageDownTradesFocusAnalysisBlock"}

\item{trade_data}{object of class "AverageDownTradesData"}
}
\value{
\code{object} object of class "AverageDownTradesFocusAnalysisBlock"
}
\description{
Public method to set trade_data slot with "AverageDownTradesData"
class object
}

