% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buys_and_sells.r
\docType{methods}
\name{setTradeDataObject,BuysAndSellsAnalysisBlock,ExtendedTradeData-method}
\alias{setTradeDataObject,BuysAndSellsAnalysisBlock,ExtendedTradeData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{BuysAndSellsAnalysisBlock,ExtendedTradeData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "BuysAndSellsAnalysisBlock"}

\item{trade_data}{object of class "ExtendedTradeData"}
}
\value{
\code{object} object of class "BuysAndSellsAnalysisBlock"
}
\description{
Public method to set trade_data slot with "ExtendedTradeData"
class object
}

