% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trades_summary.r
\docType{methods}
\name{setTradeDataObject,ExtendedTradesSummaryAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,ExtendedTradesSummaryAnalysisBlock,TradeData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{ExtendedTradesSummaryAnalysisBlock,TradeData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "ExtendedTradesSummaryAnalysisBlock"}

\item{trade_data}{object of class "TradeData"}
}
\value{
\code{object} object of class "ExtendedTradesSummaryAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradeData"
class object
}

