% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_return.r
\docType{methods}
\name{setTradeDataObject,MarketReturnAnalysisBlock,ExtendedTradeData-method}
\alias{setTradeDataObject,MarketReturnAnalysisBlock,ExtendedTradeData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{MarketReturnAnalysisBlock,ExtendedTradeData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "MarketReturnAnalysisBlock"}

\item{trade_data}{object of class "ExtendedTradeData"}
}
\value{
\code{object} object of class "MarketReturnAnalysisBlock"
}
\description{
Public method to set trade_data slot with "ExtendedTradeData"
class object
}

