% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl_traded_in_long_short_hedge.r
\docType{methods}
\name{setTradeDataObject,PnLTradedInLongShortHedgeAnalysisBlock,ValueTradedData-method}
\alias{setTradeDataObject,PnLTradedInLongShortHedgeAnalysisBlock,ValueTradedData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{PnLTradedInLongShortHedgeAnalysisBlock,ValueTradedData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "PnLTradedInLongShortHedgeAnalysisBlock"}

\item{trade_data}{object of class "TradeData"}
}
\value{
\code{object} object of class "PnLTradedInLongShortHedgeAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradeData"
class object
}

