% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_revisits.r
\docType{methods}
\name{setTradeDataObject,PositionRevisitsAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,PositionRevisitsAnalysisBlock,TradeData-method}
\title{Set trade_data object in object slot}
\usage{
\S4method{setTradeDataObject}{PositionRevisitsAnalysisBlock,TradeData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "PositionRevisitsAnalysisBlock"}

\item{trade_data}{object of class "TradeData"}
}
\value{
\code{object} object of class "PositionRevisitsAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradeData"
class object
}

