% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_revisits_delta_prev_quarter.r
\docType{methods}
\name{setTradeDataObject,PositionRevisitsDeltaPrevQuarterAnalysisBlock,VirtualTradeData-method}
\alias{setTradeDataObject,PositionRevisitsDeltaPrevQuarterAnalysisBlock,VirtualTradeData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{PositionRevisitsDeltaPrevQuarterAnalysisBlock,VirtualTradeData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "PositionRevisitsDeltaPrevQuarterAnalysisBlock"}

\item{trade_data}{object of class "VirtualTradeData"}
}
\value{
\code{object} object of class "PositionRevisitsDeltaPrevQuarterAnalysisBlock"
}
\description{
Public method to set trade_data slot with "VirtualTradeData"
class object
}

