% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_day_zero_pnl.r
\docType{methods}
\name{setTradeDataObject,PositionsHoldingDayZeroPnLAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,PositionsHoldingDayZeroPnLAnalysisBlock,TradeData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{PositionsHoldingDayZeroPnLAnalysisBlock,TradeData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "PositionsHoldingDayZeroPnLAnalysisBlock"}

\item{trade_data}{object of class "TradeData"}
}
\value{
\code{object} object of class "PositionsHoldingDayZeroPnLAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradeData"
class object
}

