% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown.r
\docType{methods}
\name{setTradeDataObject,StrategyBreakdownAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,StrategyBreakdownAnalysisBlock,TradeData-method}
\title{Set trade_data object in object slot}
\usage{
\S4method{setTradeDataObject}{StrategyBreakdownAnalysisBlock,TradeData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "StrategyBreakdownAnalysisBlock"}

\item{trade_data}{object of class "TradeData"}
}
\value{
\code{object} object of class "StrategyBreakdownAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradeData"
class object
}

