% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_pnl_and_turnover_by_event.r
\docType{methods}
\name{setTradeDataObject,StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock,TradedSignalsData-method}
\alias{setTradeDataObject,StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock,TradedSignalsData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock,TradedSignalsData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock"}

\item{trade_data}{object of class "TradedSignalsData"}
}
\value{
\code{object} object of class "StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradedSignalsData"
class object
}

