% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_pnl_on_trade_day_per_signal.r
\docType{methods}
\name{setTradeDataObject,StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock,TradedSignalsData-method}
\alias{setTradeDataObject,StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock,TradedSignalsData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock,TradedSignalsData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock"}

\item{trade_data}{object of class "TradedSignalsData"}
}
\value{
\code{object} object of class "StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradedSignalsData"
class object
}

