% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_signal_characteristic_and_effeciveness.r
\docType{methods}
\name{setTradeDataObject,StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock,TradedSignalsData-method}
\alias{setTradeDataObject,StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock,TradedSignalsData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock,TradedSignalsData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock"}

\item{trade_data}{object of class "TradedSignalsData"}
}
\value{
\code{object} object of class "StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradedSignalsData"
class object
}

