% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_value_traded_per_signal.r
\docType{methods}
\name{setTradeDataObject,StrategyBreakdownValueTradedPerSignalAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,StrategyBreakdownValueTradedPerSignalAnalysisBlock,TradeData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{StrategyBreakdownValueTradedPerSignalAnalysisBlock,TradeData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "StrategyBreakdownValueTradedPerSignalAnalysisBlock"}

\item{trade_data}{object of class "TradeData"}
}
\value{
\code{object} object of class "StrategyBreakdownValueTradedPerSignalAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradeData"
class object
}

