% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades_performance_on_results_day.r
\docType{methods}
\name{setTradeDataObject,TradesPerformanceOnResultsDayAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,TradesPerformanceOnResultsDayAnalysisBlock,TradeData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{TradesPerformanceOnResultsDayAnalysisBlock,TradeData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "TradesPerformanceOnResultsDayAnalysisBlock"}

\item{trade_data}{object of class "TradeData"}
}
\value{
\code{object} object of class "TradesPerformanceOnResultsDayAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradeData"
class object
}

