% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades_pnl_outof_on_results_day.r
\docType{methods}
\name{setTradeDataObject,TradesPnLOutOfOnResultsDayAnalysisBlock,TradesOnResultsDayData-method}
\alias{setTradeDataObject,TradesPnLOutOfOnResultsDayAnalysisBlock,TradesOnResultsDayData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{TradesPnLOutOfOnResultsDayAnalysisBlock,TradesOnResultsDayData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "TradesPnLOutOfOnResultsDayAnalysisBlock"}

\item{trade_data}{object of class "TradesOnResultsDayData"}
}
\value{
\code{object} object of class "TradesPnLOutOfOnResultsDayAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradesOnResultsDayData"
class object
}

