% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_traded_in_long_short_hedge.r
\docType{methods}
\name{setTradeDataObject,ValueTradedInLongShortHedgeAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,ValueTradedInLongShortHedgeAnalysisBlock,TradeData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{ValueTradedInLongShortHedgeAnalysisBlock,TradeData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class "ValueTradedInLongShortHedgeAnalysisBlock"}

\item{trade_data}{object of class "TradeData"}
}
\value{
\code{object} object of class "ValueTradedInLongShortHedgeAnalysisBlock"
}
\description{
Public method to set trade_data slot with "TradeData"
class object
}

