% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r, R/average_down_trades_focus.r, R/buys_and_sells.r, R/extended_trades_summary.r, R/market_return.r, R/value_traded_in_long_short_hedge.r, R/pnl_traded_in_long_short_hedge.r, R/position_revisits.r, R/position_revisits_delta_prev_quarter.r, R/positions_holding_day_zero_pnl.r, R/relative_market_return.r, R/strategy_breakdown.r, R/strategy_breakdown_value_traded_per_signal.r, R/strategy_breakdown_pnl_and_turnover_by_event.r, R/strategy_breakdown_pnl_on_trade_day_per_signal.r, R/strategy_breakdown_signal_characteristic_and_effeciveness.r, R/trades_performance_on_results_day.r, R/trades_pnl_outof_on_results_day.r
\docType{methods}
\name{setTradeDataObject}
\alias{setTradeDataObject}
\alias{setTradeDataObject,AverageDownTradesFocusAnalysisBlock,AverageDownTradesData-method}
\alias{setTradeDataObject,BuysAndSellsAnalysisBlock,ExtendedTradeData-method}
\alias{setTradeDataObject,ExtendedTradesSummaryAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,MarketReturnAnalysisBlock,ExtendedTradeData-method}
\alias{setTradeDataObject,PnLTradedInLongShortHedgeAnalysisBlock,ValueTradedData-method}
\alias{setTradeDataObject,PositionRevisitsAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,PositionRevisitsDeltaPrevQuarterAnalysisBlock,VirtualTradeData-method}
\alias{setTradeDataObject,PositionsHoldingDayZeroPnLAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,RelativeMarketReturnAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,StrategyBreakdownAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock,TradedSignalsData-method}
\alias{setTradeDataObject,StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock,TradedSignalsData-method}
\alias{setTradeDataObject,StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock,TradedSignalsData-method}
\alias{setTradeDataObject,StrategyBreakdownValueTradedPerSignalAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,TradesPerformanceOnResultsDayAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,TradesPnLOutOfOnResultsDayAnalysisBlock,TradesOnResultsDayData-method}
\alias{setTradeDataObject,ValueTradedInLongShortHedgeAnalysisBlock,TradeData-method}
\alias{setTradeDataObject,VirtualAnalysisBlock,VirtualTradeData-method}
\title{Set trade_data object in object slot}
\usage{

  \S4method{setTradeDataObject}{AverageDownTradesFocusAnalysisBlock,AverageDownTradesData}(object,
  trade_data)


  \S4method{setTradeDataObject}{BuysAndSellsAnalysisBlock,ExtendedTradeData}(object,
  trade_data)


  \S4method{setTradeDataObject}{ExtendedTradesSummaryAnalysisBlock,TradeData}(object,
  trade_data)


  \S4method{setTradeDataObject}{MarketReturnAnalysisBlock,ExtendedTradeData}(object,
  trade_data)


  \S4method{setTradeDataObject}{ValueTradedInLongShortHedgeAnalysisBlock,TradeData}(object,
  trade_data)


  \S4method{setTradeDataObject}{PnLTradedInLongShortHedgeAnalysisBlock,ValueTradedData}(object,
  trade_data)

\S4method{setTradeDataObject}{PositionRevisitsAnalysisBlock,TradeData}(object,
  trade_data)


  \S4method{setTradeDataObject}{PositionRevisitsDeltaPrevQuarterAnalysisBlock,VirtualTradeData}(object,
  trade_data)


  \S4method{setTradeDataObject}{PositionsHoldingDayZeroPnLAnalysisBlock,TradeData}(object,
  trade_data)


  \S4method{setTradeDataObject}{RelativeMarketReturnAnalysisBlock,TradeData}(object,
  trade_data)

\S4method{setTradeDataObject}{StrategyBreakdownAnalysisBlock,TradeData}(object,
  trade_data)


  \S4method{setTradeDataObject}{StrategyBreakdownValueTradedPerSignalAnalysisBlock,TradeData}(object,
  trade_data)


  \S4method{setTradeDataObject}{StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock,TradedSignalsData}(object,
  trade_data)


  \S4method{setTradeDataObject}{StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock,TradedSignalsData}(object,
  trade_data)


  \S4method{setTradeDataObject}{StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock,TradedSignalsData}(object,
  trade_data)


  \S4method{setTradeDataObject}{TradesPerformanceOnResultsDayAnalysisBlock,TradeData}(object,
  trade_data)


  \S4method{setTradeDataObject}{TradesPnLOutOfOnResultsDayAnalysisBlock,TradesOnResultsDayData}(object,
  trade_data)
}
\arguments{
\item{object}{object of class derived from"VirtualAnalysisBlock"}

\item{trade_data}{object of class derirved from "VirtualTradeData"}
}
\value{
\code{object} object of class derived from "VirtualAnalysisBlock"

\code{object} object of class "PnLTradedInLongShortHedgeAnalysisBlock"
}
\description{
Public method to set trade_data slot with object of class derirved from "VirtualTradeData"
}
\section{Methods (by class)}{
\itemize{
\item \code{object = AverageDownTradesFocusAnalysisBlock,trade_data = AverageDownTradesData}: Set trade_data object in object slot

\item \code{object = BuysAndSellsAnalysisBlock,trade_data = ExtendedTradeData}: Set trade_data object in object slot

\item \code{object = ExtendedTradesSummaryAnalysisBlock,trade_data = TradeData}: Set trade_data object in object slot

\item \code{object = MarketReturnAnalysisBlock,trade_data = ExtendedTradeData}: Set trade_data object in object slot

\item \code{object = ValueTradedInLongShortHedgeAnalysisBlock,trade_data = TradeData}: Set trade_data object in object slot

\item \code{object = PnLTradedInLongShortHedgeAnalysisBlock,trade_data = ValueTradedData}: Set trade_data object in object slot

\item \code{object = PositionRevisitsAnalysisBlock,trade_data = TradeData}: Set trade_data object in object slot

\item \code{object = PositionRevisitsDeltaPrevQuarterAnalysisBlock,trade_data = VirtualTradeData}: Set trade_data object in object slot

\item \code{object = PositionsHoldingDayZeroPnLAnalysisBlock,trade_data = TradeData}: Set trade_data object in object slot

\item \code{object = RelativeMarketReturnAnalysisBlock,trade_data = TradeData}: Set trade_data object in object slot

\item \code{object = StrategyBreakdownAnalysisBlock,trade_data = TradeData}: Set trade_data object in object slot

\item \code{object = StrategyBreakdownValueTradedPerSignalAnalysisBlock,trade_data = TradeData}: Set trade_data object in object slot

\item \code{object = StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock,trade_data = TradedSignalsData}: Set trade_data object in object slot

\item \code{object = StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock,trade_data = TradedSignalsData}: Set trade_data object in object slot

\item \code{object = StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock,trade_data = TradedSignalsData}: Set trade_data object in object slot

\item \code{object = TradesPerformanceOnResultsDayAnalysisBlock,trade_data = TradeData}: Set trade_data object in object slot

\item \code{object = TradesPnLOutOfOnResultsDayAnalysisBlock,trade_data = TradesOnResultsDayData}: Set trade_data object in object slot
}}

