% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_value_traded_per_signal.r
\docType{data}
\name{strategy_breakdown_per_signal_base_cols}
\alias{strategy_breakdown_per_signal_base_cols}
\title{List of traded signals data base column names}
\format{An object of class \code{character} of length 7.}
\usage{
strategy_breakdown_per_signal_base_cols
}
\description{
List of traded signals data base column names
}
\keyword{datasets}

