% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_value_traded_per_signal.r
\docType{data}
\name{strategy_breakdown_per_signal_signal_cols}
\alias{strategy_breakdown_per_signal_signal_cols}
\title{List of traded signals names}
\format{An object of class \code{character} of length 25.}
\usage{
strategy_breakdown_per_signal_signal_cols
}
\description{
List of traded signals names
}
\keyword{datasets}

