% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown.r
\docType{class}
\name{StrategyBreakdownAnalysisBlock-class}
\alias{StrategyBreakdownAnalysisBlock-class}
\title{Analysis Module for computation of per strategy breakdown}
\description{
Computation block class to pull data required for Computation of strategy breakdown plot.
Pulls data required for computation and adds required columns.
Generates multifacet strategy breakdown ggplot.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler"
}

