% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_aum_and_turnover.r
\docType{class}
\name{StrategyBreakdownAUMAndTurnoverAnalysisBlock-class}
\alias{StrategyBreakdownAUMAndTurnoverAnalysisBlock-class}
\title{Analysis Module for computation of per strategy breakdown AUM and Turnover}
\description{
Computation block class that computes Assets Under Management value
and Turnover per strategy. Depends on data generated by.
StrategyBreakdownAnalysisBlock. Generates ggplot with AUM and Turnover
per strategy.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualStrategyDataHandler"
}

