% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown.r
\docType{class}
\name{StrategyBreakDownData-class}
\alias{StrategyBreakDownData-class}
\title{StrategyBreakDownData Reference Data class.}
\description{
Concrete S4 class storing data with per strategy breakdown of
Portfolio quantities
}
\details{
Inherits from "VirtualStrategyData"
}

