% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_pnl_and_turnover_by_event.r
\docType{class}
\name{StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock-class}
\alias{StrategyBreakdownPnLAndTurnoverPerEventAnalysisBlock-class}
\title{Analysis Module for computation of per signal breakdown AUM and Turnover}
\description{
Computation block class that computes Assets Under Management value
and Turnover per signal. Depends on data generated by
StrategyBreakdownValueTradedPerSignalAnalysisBlock.
Generates ggplot with AUM and Turnover per signal.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler"
}

