% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_signal_characteristic_and_effeciveness.r
\docType{class}
\name{StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock-class}
\alias{StrategyBreakdownSignalCharacteristicAndEffectivenessAnalysisBlock-class}
\title{Analysis Module for computation of per signal characteristic
and effectiveness}
\description{
Computation block class that computes characteristics of traded signals
and its effectiveness.Depends on data generated by
StrategyBreakdownAnalysisBlock. Generates ggplot with signal attributes.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler"
}

