% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_total_pnl_and_pnl_delta.r
\docType{class}
\name{StrategyBreakdownTotalAndDeltaPnLAnalysisBlock-class}
\alias{StrategyBreakdownTotalAndDeltaPnLAnalysisBlock-class}
\title{Analysis Module for computation of Total and Delta PnL per strategy}
\description{
Computation block class that computes total and delta Profit and Loss.
Depends on output from StrategyBreakdownAnalysisBlock.
Generates ggplot computed data.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualStrategyDataHandler"
}

