% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_value_traded_per_signal.r
\docType{class}
\name{StrategyBreakdownValueTradedPerSignalAnalysisBlock-class}
\alias{StrategyBreakdownValueTradedPerSignalAnalysisBlock-class}
\title{Analysis Module for computation of breakdown of portfolio
attributes per traded signal}
\description{
Computation block class to pull data required for Computation of
traded signals breakdown plot.
Pulls data required for computation and adds required columns.
Generates multifacet signal breakdown ggplot.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler",
              "VirtualEventDataHandler"
}

