% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.AnalysisClasses.r
\docType{package}
\name{TE.AnalysisClasses}
\alias{TE.AnalysisClasses}
\alias{TE.AnalysisClasses-package}
\title{TE.AnalysisClasses: Trading Enhancement Data access, cashing and preprocessing}
\description{
The TE.AnalysisClasses package providing classes for embedding of computation of
various analysis data and visualisation via ggplot
}

