% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_value_traded_per_signal.r
\docType{class}
\name{TradedSignalsData-class}
\alias{TradedSignalsData-class}
\title{TradedSignalsData Reference Data class.}
\description{
Concrete S4 class storing data of signals
of stock traded within given timeframe
}
\details{
Inherits from "VirtualTradeData"
}

