% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_return.r
\docType{class}
\name{TradesExtendedReturnPerMonth-class}
\alias{TradesExtendedReturnPerMonth-class}
\title{TradesExtendedReturnPerMonth Reference Data class.}
\description{
Concrete S4 class storing data of Extended Trades Return per Month
Generated only by MarketReturnAnalysisBlock module.
}
\details{
Inherits from "VirtualTradeData"
}

