% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades_performance_on_results_day.r
\docType{class}
\name{TradesOnResulsData-class}
\alias{TradesOnResulsData-class}
\title{Trades on Results Reference Data class.,}
\description{
Concrete S4 class storing data trades happening on Results day
}
\details{
Inherits from "VirtualTradeData"
}

