% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades_performance_on_results_day.r
\docType{class}
\name{TradesPerformanceOnResultsDayAnalysisBlock-class}
\alias{TradesPerformanceOnResultsDayAnalysisBlock-class}
\title{Analysis Module for extraction of performance of trades on results day}
\description{
Computation block class compute data for
long and short time performance of trades done on result day.
Specificaly it looks on performance of stock that was
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler",
              "VirtualPriceDataHandler"
}

