% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades_pnl_outof_on_results_day.r
\docType{class}
\name{TradesPnLOutOfOnResultsDayAnalysisBlock-class}
\alias{TradesPnLOutOfOnResultsDayAnalysisBlock-class}
\title{Analysis Module for extraction of performance of trades on results day}
\description{
Computation block class compute data for
long and short time performance of trades done on result day.
Specificaly it looks on performance of stock that was
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler",
              "VirtualPriceDataHandler"
}

