% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_traded_in_long_short_hedge.r
\docType{class}
\name{ValueTradedData-class}
\alias{ValueTradedData-class}
\title{ValueTradedData Reference Data class.}
\description{
Concrete S4 class storing data of traded value data.
}
\details{
Inherits from "VirtualTradeData"
}

