% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_traded_in_long_short_hedge.r
\docType{class}
\name{ValueTradedInLongShortHedgeAnalysisBlock-class}
\alias{ValueTradedInLongShortHedgeAnalysisBlock-class}
\title{Analysis Module for Computation of traded value data.}
\description{
Computation block class to pull data required for Computation of
Traded Value categorized per long , short and hedge strats
}
\details{
Pulls data required for computation and adds required columns.

Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler"
}

