% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\docType{class}
\name{VirtualAggregateAnalysisBlock-class}
\alias{VirtualAggregateAnalysisBlock-class}
\title{Block that holds only ggplot data and ggplot that has been gathered from
other blocks and aggregated}
\description{
Inherits from "VirtualAnalysisBlock",
}
\section{Slots}{

\describe{
\item{\code{ggplot_data_aggregator}}{object of class "function"}

\item{\code{ggplot_object_modifier}}{object of class "function"}

\item{\code{ggplot_format_modifier}}{object of class "function"}

\item{\code{ui_options_modifier}}{object of class "function"}
}}

