% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\docType{class}
\name{VirtualAnalysisBlock-class}
\alias{VirtualAnalysisBlock-class}
\title{Virtual S4 class implementing handling of analysis objects.}
\description{
Class that is to be inherited by any objects
that will encapsulate specific analysis computations.
}
\details{
Inherits from "VirtualDataSourceClient"
}
\section{Slots}{

\describe{
\item{\code{ggplot}}{object of class "ggplot"}

\item{\code{ggplot_data}}{"data.frame",}

\item{\code{frontend_data}}{"data.frame",}

\item{\code{output}}{"NullableReferenceData"}
}}

