% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\docType{methods}
\name{aggregateAnalysisData}
\alias{aggregateAnalysisData}
\alias{aggregateAnalysisData,VirtualAggregateAnalysisBlock,data.frame,data.frame-method}
\title{Aggregate data into the block}
\usage{
aggregateAnalysisData(object, data, with_id)


  \S4method{aggregateAnalysisData}{VirtualAggregateAnalysisBlock,data.frame,data.frame}(object,
  data, with_id)
}
\arguments{
\item{object}{object of class "VirtualAggregateAnalysisBlock"}

\item{data}{data of class "data.frame"}

\item{with_id}{with_id of class "data.frame"}
}
\value{
\code{object} object of class "VirtualAggregateAnalysisBlock"
}
\description{
Aggregate data into the block
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualAggregateAnalysisBlock,data = data.frame,with_id = data.frame}: Aggregate data into the block
}}

