% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades.r
\docType{class}
\name{AverageDownTradesAnalysisBlock-class}
\alias{AverageDownTradesAnalysisBlock-class}
\title{Analysis Module for computation of Market Offside Positions}
\description{
Computation block class to pull data required for Computation of offside positions
Pulls data required for computation and adds required columns.
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualPositionDataHandler",
              "VirtualTradeDataHandler"
}

