% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\docType{methods}
\name{getOutputFrontendData}
\alias{getOutputFrontendData}
\alias{getOutputFrontendData,VirtualAnalysisBlock-method}
\title{Returns generated output frontend data}
\usage{
getOutputFrontendData(object)

\S4method{getOutputFrontendData}{VirtualAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}
}
\value{
\code{frontend_data} object of class "data.frame"

\code{frontend_data} object of class "data.frame"
}
\description{
Returns data to be used by any frontend using the module
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualAnalysisBlock}: Returns generated output frontend data

Returns data to be used by any frontend using the module
}}

