% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\docType{methods}
\name{getOutputGGPlot}
\alias{getOutputGGPlot}
\alias{getOutputGGPlot,VirtualAnalysisBlock-method}
\title{Returns generated output ggplot}
\usage{
getOutputGGPlot(object)

\S4method{getOutputGGPlot}{VirtualAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}
}
\value{
\code{ggplot} object of class "ggplot"

\code{ggplot} object of class "ggplot"
}
\description{
Returns ggplot object generated as a result of analysis computation.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualAnalysisBlock}: Returns generated output ggplot

Returns ggplot object generated as a result of analysis computation.
}}

