% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\name{getOutputGGPlotData}
\alias{getOutputGGPlotData}
\title{Returns generated output ggplot data}
\usage{
getOutputGGPlotData(object, ...)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}
}
\value{
\code{ggplot_data} object of class "data.frame"
}
\description{
Returns "data.frame" with data used to generate ggplot
}

