% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\docType{methods}
\name{getOutputObject}
\alias{getOutputObject}
\alias{getOutputObject,VirtualAnalysisBlock-method}
\title{Returns generated output object}
\usage{
getOutputObject(object)

\S4method{getOutputObject}{VirtualAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}
}
\value{
\code{output} object of class derived from "VirtualReferenceData"
}
\description{
Returns Reference Data Object that possibly was computed by
analysis object. If module doesn't compute any output
returns NULL
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualAnalysisBlock}: Returns generated output object

Returns Reference Data Object that possibly was computed by
analysis object. If module doesn't compute any output
returns NULL
}}

