% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\name{modifyGGPlotFormat}
\alias{modifyGGPlotFormat}
\title{Modify the plot format: i.e. add gg functions the plot stream}
\usage{
modifyGGPlotFormat(object)
}
\arguments{
\item{object}{object of class "VirtualAggregateAnalysisBlock"}
}
\value{
\code{object} object of class "VirtualAggregateAnalysisBlock"
}
\description{
Modify the plot format: i.e. add gg functions the plot stream
}

