% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\name{modifyGGPlotObject}
\alias{modifyGGPlotObject}
\title{Modify the plot object: i.e. reassign to fields internal to
the ggplot}
\usage{
modifyGGPlotObject(object)
}
\arguments{
\item{object}{object of class "VirtualAggregateAnalysisBlock"}
}
\value{
\code{object} object of class "VirtualAggregateAnalysisBlock"
}
\description{
Modify the plot object: i.e. reassign to fields internal to
the ggplot
}

