% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_analysis_block.r
\name{modifyUIOptions}
\alias{modifyUIOptions}
\title{Modify the ui options list}
\usage{
modifyUIOptions(object)
}
\arguments{
\item{object}{object of class "VirtualAggregateAnalysisBlock"}
}
\value{
\code{object} object of class "VirtualAggregateAnalysisBlock"
}
\description{
Modify the ui options list
}

