% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_revisits.r
\docType{methods}
\name{Process,PositionRevisitsAnalysisBlock-method}
\alias{Process,PositionRevisitsAnalysisBlock-method}
\title{Trigger computation of analysis data.}
\usage{
\S4method{Process}{PositionRevisitsAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "StrategyBreakdownTotalAndDeltaPnLAnalysisBlock"}
}
\value{
\code{object} object object of class "StrategyBreakdownTotalAndDeltaPnLAnalysisBlock"
}
\description{
Trigger computation of analysis data.
}

