% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\name{Process}
\alias{Process}
\title{Trigger computation of analysis data.}
\usage{
Process(object)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}
}
\description{
Trigger computation of analysis data.
}

