% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_block.r
\name{Process}
\alias{Process}
\title{Trigger computation of analysis data.}
\usage{
Process(object, ...)
}
\arguments{
\item{object}{object of class "VirtualAnalysisBlock"}
}
\value{
\code{object} object object of class "VirtualAnalysisBlock"
}
\description{
Private method to set slot with data used by frontend utility.
Needs to be implemented by each class derived from "VirtualAnalysisBlock"
}

