% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_instrument_marginal_contribution_to_risk.r
\docType{methods}
\name{setPortfolioDataObject,IndexPortfolioInstrumentMCTRAnalysisBlock,VirtualIndexPortfolio-method}
\alias{setPortfolioDataObject,IndexPortfolioInstrumentMCTRAnalysisBlock,VirtualIndexPortfolio-method}
\title{Set portfolio object in object slot}
\usage{

  \S4method{setPortfolioDataObject}{IndexPortfolioInstrumentMCTRAnalysisBlock,VirtualIndexPortfolio}(object,
  portfolio)
}
\arguments{
\item{object}{object of class "IndexPortfolioInstrumentMCTRAnalysisBlock"}

\item{portfolio}{object of class "VirtualIndexPortfolio"}
}
\value{
\code{object} object of class "IndexPortfolioInstrumentMCTRAnalysisBlock""
}
\description{
Public method to set portfolio slot with "VirtualIndexPortfolio"
class object
}

