% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_breakdown_pnl_on_trade_day_per_signal.r
\docType{class}
\name{StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock-class}
\alias{StrategyBreakdownPnLOnTradeDayPerSignalAnalysisBlock-class}
\title{Analysis Module for computation of per Signal breakdown PnL}
\description{
Computation block class that computes Profit and Loss value
per signal. Depends on data generated by
StrategyBreakdownAnalysisBlock. Generates ggplot with PnL
per signal
}
\details{
Inherits from "VirtualAnalysisBlock",
              "VirtualTradeDataHandler"
}

